% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{add.facets.snpR.data}
\alias{add.facets.snpR.data}
\title{Add facets to snpRdata objects}
\usage{
add.facets.snpR.data(x, facets = NULL)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{character. Facets to use.}
}
\description{
Adds facets to snpRdata objects, following the rules described in
\code{\link{Facets_in_snpR}}. Internal, called by many other functions when
facets are requested.
}
