% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{calc_association}
\alias{calc_association}
\title{Run case/control association tests on SNP data.}
\usage{
calc_association(x, facets = NULL, response, method = "armitage",
  w = c(0, 1, 2))
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{character, default NULL. Categorical metadata variables by
which to break up analysis. See \code{\link{Facets_in_snpR}} for more
details.}

\item{response}{character. Name of the column containing the response
variable of interest. Must match a column name in sample metadata. Response
must be categorical, with only two categories.}

\item{method}{character, default "armitage". Specifies association method.
Options: \itemize{ \item{armitage: } Armitage association test, based on
Armitage (1955). \item{odds_ratio: } Log odds ratio test. \item{chisq: }
Chi-squared test. } See description for more details.}

\item{w}{numeric, default c(0, 1, 2). Weight variable for each genotype for the Armitage association method. See description for details.}
}
\value{
A snpRdata object with the resulting association test results merged into the stats socket.
}
\description{
Runs several different association tests on SNP data. The response variable must
have only two different categories (as in case/control) tests. Support for more categories
or continuous data will be added later. Tests may be broken up by sample-specific facets.
}
\details{
Several methods can be used: Armitage, chi-squared, and odds ratio. For The Armitage approach
weights should be provided to the "w" argument, which specifies the weight for each possible genotype
(homozygote 1, heterozygote, homozygote 2). The default, c(0,1,2), specifies an addative model.

Continuous or multi-category data is currently not supported, but is under development.

Facets are specified as described in \code{\link{Facets_in_snpR}}. NULL and "all" facet specifications
function as described.
}
\examples{
  # add a dummy phenotype
  sample.meta <- cbind(stickSNPs@sample.meta, phenotype = sample(c("A", "B"), nrow(stickSNPs@sample.meta), T))
  x <- import.snpR.data(as.data.frame(stickSNPs), stickSNPs@snp.meta, sample.meta)
  calc_association(x, facets = c("pop", "pop.fam"), response = "phenotype", method = "armitage")

}
\references{
Armitage (1955). Tests for Linear Trends in Proportions and Frequencies. \emph{Biometrics}.
}
\author{
William Hemstrom

Keming Su

Avani Chitre
}
