% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_basic_snp_stats}
\alias{calc_basic_snp_stats}
\title{Calculate basic SNP statistics}
\usage{
calc_basic_snp_stats(
  x,
  facets = NULL,
  fst.method = "WC",
  sigma = NULL,
  step = NULL,
  par = FALSE,
  nk = TRUE
)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character. Categorical metadata variables by which to break up
analysis. See \code{\link{Facets_in_snpR}} for more details.}

\item{fst.method}{character, default "WC". Defines the FST estimator to use.
Options:
\itemize{
\item WC: Weir and Cockerham (1984).
\item Weir: Weir
(1990)
\item Hohenlohe: Hohenlohe et al (2010), identical to the STACKS
package.
\item Genepop: Rousset (2008), uses the genepop package.
}}

\item{sigma}{numeric. Designates the width of windows in kilobases. Full
window size is 6*sigma.}

\item{step}{numeric or NULL, default NULL. Designates the number of kilobases
between each window centroid. If NULL, windows are centered on each SNP.}

\item{par}{numeric or FALSE, default FALSE. If numeric, the number of cores to
use for parallel processing.}

\item{nk}{logical, default TRUE. If TRUE, weights SNP contribution to window
averages by the number of observations at those SNPs.}
}
\value{
A snpRdata object with all of the described statistics merged into the
appropriate sockets.
}
\description{
Automatically calculate most basic statistics from snpRdata. Calculates maf,
pi, ho, he, pairwise Fst, Fis, HWE divergence, finds private alleles, and uses Gaussian
smoothing to produce per-window averages of all of these.
}
\details{
The data can be broken up categorically by sample or SNP metadata, as
described in \code{\link{Facets_in_snpR}}. Note that Fst and private allele
calculations require a sample specific contrast (the pairwise part of pairwise
Fst), and thus will not be calculated unless a facet with a sample meta data
variable included is specified. The other stats, in contrast, are snp-specific
and thus ignore any snp meta data variables included in facets. Providing NULL
or "all" to the facets argument works as described in
\code{\link{Facets_in_snpR}}.
}
\examples{
x <- calc_basic_snp_stats(stickSNPs, "pop")
get.snpR.stats(x, "pop", stats = "single") # view basic stats
get.snpR.stats(x, "pop", stats = "fst") # view fst


}
\seealso{
calc_single_stats calc_pairwise_fst calc_smoothed_averages
}
\author{
William Hemstrom
}
