% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_CLD}
\alias{calc_CLD}
\title{Calculate Burrow's composite LD. Internal.}
\usage{
calc_CLD(x, facets = NULL, par = FALSE)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{facets to run}

\item{par}{number of parallel cores}
}
\description{
Called in \code{\link{calc_pairwise_ld}}. Complex function purely because the output needs to be
in the same format as that function's and to support faceting without excessive recalculation.
}
\author{
William Hemstrom
}
