% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs_functions.R
\name{calc_directionality}
\alias{calc_directionality}
\title{Calculate the directionality index from a 2d site frequency spectrum.}
\usage{
calc_directionality(sfs)
}
\arguments{
\item{sfs}{numeric matrix. A 2d site frequency spectra stored in a matrix,
with an additional "pops" attribute containing population IDs, such as
c("POP1", "POP2"), where the first pop corresponds to matrix columns and
the second to matrix rows. These objects can be produced from a dadi input
file using \code{\link{make_SFS}}.}
}
\description{
Calculates the directionality index based on a 2d SFS according to Peter and
Slatkin (2013). Input spectra can be created using the \code{\link{make_SFS}}
function. Spectra must be not be folded
}
\details{
Essentially, the directionality index measures the difference in derived
allele frequency between two populations to determine the directionality of
population spread between the two. Since the "destination" population is
sourced from but experienced more genetic drift than the "source" population,
it should have relativley more high-frequency derived alleles \emph{after the
removal of fixed ancestral alleles}. See Peter and Slatkin (2013) for
details.
}
\references{
Peter, B. M., & Slatkin, M. (2013). Detecting range expansions
  from genetic data. \emph{Evolution}, 67(11), 3274-3289.
}
