% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_genetic_distances}
\alias{calc_genetic_distances}
\title{Caculate genetic distances between groups of samples.}
\usage{
calc_genetic_distances(x, facets = NULL, method = "Edwards")
}
\arguments{
\item{x}{Either a snpRdata object or a nested, named list of allele
frequencies as produced by \code{\link{get_allele_frequencies}}.}

\item{facets}{character or NULL, default NULL. Facets for which to calculate
genetic distances, as described in \code{\link{Facets_in_snpR}}. Note that
any solely snp specific facets will get dropped, since distances are
calculated between groups of samples.}

\item{method}{character, default "Edwards". Name of the method to use.
Options: \itemize{\item{Edwards} Angular distance as described in Edwards
1971.} See details.}
}
\value{
A named, nested list containing distance measures named according to facets and facet levels.
}
\description{
Calculates the genetic distances between any sample specific facets, broken
apart by any requested snp level facets. For details on methods, see details.
Note that this particular function is not overwrite-safe.
}
\details{
Available methods: \itemize{\item{Edwards} Angular distance as described in
Edwards 1971.}

All methods first calculate allele frequency matrices using
\code{\link{get_allele_frequencies}}, then use these matrices (hereafter x)
to calculate genetic distance.

Method details: \itemize{\item{Edwards 1971:} x <- sqrt(x); x <- x%*%t(x); x
<- 1/(x/number.loci); diag(x) <- 0; x <- sqrt(x)}
}
\references{
Edwards, A. W. F. (1971). Distances between populations on the basis of gene frequencies. Biometrics, 873-881.
}
\author{
William Hemstrom
}
