% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_het_hom_ratio}
\alias{calc_het_hom_ratio}
\title{Calculate the ratio of heterozygous/homozygous sites per individual.}
\usage{
calc_het_hom_ratio(x)
}
\description{
Calculates the ratio of heterozygotes to homozygous sites across all SNPs within each individual.
Facet support (for snp level facets) needs to be added, should be straightforward.
}
