% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_hwe}
\alias{calc_hwe}
\title{Calculate HWE divergence from SNP data.}
\usage{
calc_hwe(x, facets = NULL, method = "exact")
}
\arguments{
\item{x}{Input SNP data, in the snpRdata format.}

\item{facets}{Facets to run. Non-sample level facets will be removed. Periods can be used to list complex facets, such as pop.fam to split by both pop and fam simultaniously.}
}
\value{
A snpRdata object with HWE data added.
}
\description{
\code{calc_hwe} Calculates p-values for Hardy-Weinburg Equilibrium divergence from snp.R.data objects.
}
\examples{

}
