% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_pairwise_nk}
\alias{calc_pairwise_nk}
\title{Pairwise nk from SNP data.}
\usage{
calc_pairwise_nk(x, ecs)
}
\arguments{
\item{x}{Input data frame, in allele count format as given by format_snps option 1.}

\item{ecs}{Number of extra metadata columns at the start of x *not counting the column with population IDs* but counting position, ect. as normal.}
}
\value{
A data.frame containing the number of alleles secquenced in each pair of populations.
}
\description{
\code{calc_pairwise_nk} calculates number of alleles observed in each parwise combination of populations. Wrapper function for \code{\link{calc_pairwise_Fst}}
}
\details{
Description of x:
    Must contain colums containing the number of *unique* alleles, total count of alleles sequenced in all individuals, and subsequent alleles counts for each observed allele in columns named "n_alleles", "n_total", "ni1", and "ni2". Also needs a column containing the position of each SNP, in bp. This matches the allele count/bayescan format as given by format_snps option one. Should also contain columns titled "group", "position", and "pop", which contain the linkage group/chr, position in bp, and population ID for each SNP.
}
\examples{
calc_pairwise_nk(ac, 3)
}
