% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_pi}
\alias{calc_pi}
\alias{calc_maf}
\title{Calculate standard single-SNP genetic statistics.}
\usage{
calc_pi(x, facets = NULL)
}
\arguments{
\item{x}{snpRdata. Input SNP data.}

\item{facets}{character. Facets to use.}
}
\value{
snpRdata object with containing merged in pi data in the stats socket
}
\description{
These functions calculate a basic genetic statistics from SNP data contained in snpRdata objects, splitting samples by any number of provided facets. Since these statistics all use only a single SNP at a time, they ignore any SNP specific facet levels.
}
\details{
Facet designation follows a specific format. Facets are given as a character
vector, where each entry designates one unique combination of levels over
which to seperate the data. Levels within a facet are seperated by a '.'. Each
facet can contain multiple snp and/or sample levels.

All listed facet levels must match column names in the snp of sample level facet data provided when constructing a snpRdata object with \code{\link{import.snpR.data}}. The order of the provided levels within each facet does not matter.

See examples.
}
\section{pi}{


Calculates pi (genetic diversity/average number of pairwise differences)
according to Hohenlohe et al. (2010).



For example, c("group.pop", "pop") would split the data first by both group
and pop and then by pop alone.

Since pi is a non-pairwise statistic that is independantly calculated for each
snp, snp levels in facets are ignored.
}

\examples{
# base facet
x <- calc_pi(stickSNPs)

# multiple facets
x <- calc_pi(stickSNPs, facets = c("pop", "pop.fam"))
}
