% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_pid}
\alias{calc_pid}
\title{Probability of Identity}
\usage{
calc_pid(x, facets = NULL)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{facets over which to calculate PID. See
\code{\link{Facets_in_snpR}} for details. SNP facets are ignored.}
}
\value{
A snpRdata object with PID info merged in, fetchable with
\code{\link{get.snpR.stats}}
}
\description{
Calculates the probability of sampling two individuals from the same
population with identical genotypes.
}
\details{
Both per-locus and across all locus probabilities are calculated. Two
statistics are calculated:
\enumerate{
\item PID: unbiased probability of randomly sampling two identical individuals
from a population, corrected for sample size.
\item PIDsib: probability of
randomly sampling two genetically identical siblings from a population.
}
}
\examples{
x <- calc_pid(stickSNPs, "pop")
get.snpR.stats(x, "pop", "pid")


}
\references{
Paetkau, D., Waits, L. P., Clarkson, P. L., Craighead, L., Vyse, E., Ward,
R., & Strobeck, C. (1998). Variation in Genetic Diversity across the Range of
North American Brown Bears. Conservation Biology, 12(2), 418–429.
https://doi.org/10.1111/j.1523-1739.1998.96457.x
}
\author{
William Hemstrom
}
