% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_private}
\alias{calc_private}
\title{Private Alleles from SNP data.}
\usage{
calc_private(x, facets = NULL)
}
\arguments{
\item{x}{Input data, in the format given by format_snps output option "ac". Must contain a column with pop info named "pop".}

\item{ecs}{Numeric, number of metadata columns at the start of x to maintain for output.}
}
\value{
Data frame with no metadata but sorted identically to input noting if each (row) locus is private in each (column) population, coded as 1 if private, 0 if not.
}
\description{
\code{check_private} checks for the presence of private alleles at any loci across multiple populations.
}
\details{
Description of x:
   Must contain colums containing columns containing the allele counts for each observed allele in columns named "ni1"" and "ni2". Also needs a column containing population IDs, in a column named "pop". Note that SNPs must be identically sorted in each pop, and sorted first by pop! For example, sorted by pop, group, position then position. Runs with output from format_snps output option 1.
}
\examples{
#add a private allele for demonstration purposes.
pall <- data.frame(snp = rep(10000, 6), position = rep(1, 6), group = rep("Dup", 6), pop = c("ASP", "CLF", "PAL", "OPL", "SMR", "UPD"), n_total = rep(100, 6), n_alleles = rep (2, 6), ni1 = c(rep(0, 5), 20), ni2 = c(rep(100, 5), 80))
pops <- table(substr(colnames(stickSNPs[,4:ncol(stickSNPs)]), 1, 3))
ac <- format_snps(stickSNPs, 3, pop = pops)
ac <- rbind(ac, pall)
ac <- dplyr::arrange(ac, pop, position, group)
check_private(ac)

}
