% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_seg_sites}
\alias{calc_seg_sites}
\title{Calculate the number of segregating sites.}
\usage{
calc_seg_sites(x, facets = NULL, rarefaction = TRUE)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{facets over which to calculate the number of segregating sites.
See \code{\link{Facets_in_snpR}} for details.}

\item{rarefaction}{logical, default TRUE. Should the number of segregating
sites be estimated via rarefaction? See details.}
}
\value{
A snpRdata object with seg_sites merged into the weighted.means 
  slot.
}
\description{
Calculates the segregating loci for each facet level with or without 
rarefaction to account for differing sample size and missing data levels
across populations.
}
\details{
Rarefaction is done by determining the probability of drawing at least one
copy of each allele given a standardized sample size, \emph{g} for each loci.
This is then summed across all loci to estimate the number of segregating
sites.

Note that \emph{g} will vary across loci due to differences in sequencing
coverage at those loci, equal to the smallest number of alleles sequenced in
any population at that locus minus one.
}
\examples{
# base facet
x <- calc_seg_sites(stickSNPs, rarefaction = FALSE)
get.snpR.stats(x, stats = "seg_sites")$weighted.means

# multiple facets
x <- calc_seg_sites(stickSNPs, c("pop", "fam"))
get.snpR.stats(x, c("pop", "fam"), 
               stats = "seg_sites")$weighted.means
}
\author{
William Hemstrom
}
