% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_single_stats}
\alias{calc_single_stats}
\alias{calc_pi}
\alias{calc_hwe}
\alias{calc_ho}
\alias{calc_private}
\alias{calc_maf}
\alias{calc_he}
\alias{calc_allelic_richness}
\title{Calculate standard single-SNP genetic statistics}
\usage{
calc_pi(x, facets = NULL)

calc_maf(x, facets = NULL)

calc_ho(x, facets = NULL)

calc_private(x, facets = NULL, rarefaction = TRUE)

calc_hwe(
  x,
  facets = NULL,
  method = "exact",
  fwe_method = "BY",
  fwe_case = c("by_facet", "overall")
)

calc_he(x, facets = NULL)

calc_allelic_richness(x, facets = NULL)
}
\arguments{
\item{x}{snpRdata. Input SNP data.}

\item{facets}{character. Categorical metadata variables by which to break up
analysis. See \code{\link{Facets_in_snpR}} for more details.}

\item{method}{character, default "exact". Defines the method to use for
calculating p-values for HWE divergence. Options: \itemize{ \item{exact: }
Uses the exact test as described in Wigginton et al (2005). \item{chisq: }
Uses a chi-squared test. } See details}

\item{fwe_method}{character, default "BY". Type of Family-Wise Error
correction (multiple testing correction) to use. For details and options,
see \code{\link[stats]{p.adjust}}. If no correction is desired, set this
argument to "none".}

\item{fwe_case}{character, default c("by_facet", "by_subfacet", "overall").
How should Family-Wise Error correction (multiple testing correction) be
applied? \itemize{\item{"by_facet":} Each facet supplied (such as pop or
pop.fam) is treated as a set of tests. \item{"by_subfacet":} Each level of
each subfacet is treated as a separate set of tests. \item{"overall":} All
tests are treated as a set.}}
}
\value{
snpRdata object with requested stats merged into the stats socket
}
\description{
These functions calculate a basic genetic statistics from SNP data contained
in snpRdata objects, splitting samples by any number of provided facets. Since
these statistics all use only a single SNP at a time, they ignore any SNP
specific facet levels.
}
\details{
The data can be broken up categorically by sample metadata, as described in
\code{\link{Facets_in_snpR}}.
}
\section{Functions}{
\itemize{
\item \code{calc_pi()}: \eqn{\pi} (nucleotide diversity/average number of pairwise differences)

\item \code{calc_maf()}: minor allele frequency

\item \code{calc_ho()}: observed heterozygosity

\item \code{calc_private()}: find private alleles

\item \code{calc_hwe()}: p-values for Hardy-Weinberg Equilibrium divergence

\item \code{calc_he()}: expected heterozygosity

\item \code{calc_allelic_richness()}: allelic richness (standardized number of alleles per locus via rarefaction)

}}
\section{\eqn{\pi}}{


 Calculates \eqn{\pi} (nucleotide diversity/average number of pairwise
 differences) according to Hohenlohe et al. (2010).
}

\section{\ifelse{html}{\out{H<sub>E</sub>}}{\eqn{H_E}}}{
 
 
 Calculates traditional
 expected heterozygosity \eqn{2pq}. Note that this will produce results
 almost identical to \eqn{\pi}.
}

\section{\ifelse{html}{\out{H<sub>O</sub>}}{\eqn{H_O}}}{


 Calculates observed heterozygosity.
}

\section{maf}{


 Calculates minor allele frequencies and note identities and counts of major
 and minor alleles.
}

\section{private alleles}{


 Determines if each SNP is a private allele across all levels in each sample
 facet. Will return an error if no sample facets are provided. If rarefaction
 is requested, the estimated number of private alleles will be calculated
 according to Smith and Grassle (1977). Note that the standardized sample
 size (\emph{g}) will vary across loci due to differences in sequencing
 coverage at those loci, equal to the smallest number of alleles sequenced in
 any population at that locus minus one. Instead of weighted averages, the
 value stored in the \code{$weighted.means} slot in the returned value is
 the total number of private alleles per population.
}

\section{hwe}{


 Calculates a p-value for the null hypothesis that a population is in HWE at
 a given locus. Several methods available: \itemize{ \item{"exact"} Exact
 test according to Wigginton, JE, Cutler, DJ, and Abecasis, GR (2005).
 Slightly slower. \item{"chisq"} Chi-squared test. May produce poor results
 when sample sizes for any observed or expected genotypes are small. }

 For the exact test, code derived from
 \url{http://csg.sph.umich.edu/abecasis/Exact/snp_hwe.r}
}

\section{allelic richness}{


 Calculates the allelic richness, the estimated number of alleles per locus
 standardized via rarefaction for sample size according to Hurlburt (1971).
 Note that the standardized sample size (\emph{g}) will vary across loci
 due to differences in sequencing coverage at those loci, equal to the 
 smallest number of alleles sequenced in any population at that locus minus 
 one. Weighted averages are weighted by \emph{g}.
}

\examples{
# base facet
x <- calc_pi(stickSNPs)
get.snpR.stats(x)

# multiple facets
x <- calc_pi(stickSNPs, facets = c("pop", "pop.fam"))
get.snpR.stats(x, c("pop", "pop.fam"))

# HWE with family-wise error correction
\dontrun{
x <- calc_hwe(stickSNPs, facets = c("pop", "pop.fam"))
get.snpR.stats(x, c("pop", "pop.fam"))
}
}
\references{
Wigginton, JE, Cutler, DJ, and Abecasis, GR (2005). 
 \emph{American Journal of Human Genetics}
 
 Hohenlohe et al. (2010). \emph{PLOS Genetics}.
 
 Hurlburt (1971). \emph{Ecology}.
 
 Smith and Grassle (1977). \emph{Biometrics}
}
\author{
William Hemstrom
}
