% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_functions.R
\name{calc_smoothed_averages}
\alias{calc_smoothed_averages}
\title{Gaussian smooth statistics across sliding windows.}
\usage{
calc_smoothed_averages(
  x,
  facets = NULL,
  sigma,
  step = NULL,
  nk = TRUE,
  stats.type = c("single", "pairwise"),
  par = FALSE
)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character or NULL, default NULL. Categories by which to break up
windows.}

\item{sigma}{numeric. Designates the width of windows in kilobases. Full
window size is 6*sigma.}

\item{step}{numeric or NULL, default NULL. Designates the number of kilobases
between each window centroid. If NULL, windows are centered on each SNP.}

\item{nk}{logical, default TRUE. If TRUE, weights SNP contribution to window
averages by the number of observations at those SNPs.}

\item{stats.type}{character, default c("single", "pairwise"). Designates the
statistic(s) to smooth, either "single",  "pairwise", or c("single",
"pairwise"). See details.}

\item{par}{numeric or FALSE, default FALSE. If numeric, the number of cores to
use for parallel processing.}
}
\value{
snpRdata object with smoothed averages for any requested statistics
 merged into the window.stats or pairwise.window.stats slots.
}
\description{
Calculates gaussian smoothed average values for statistics and the genomic
position at which it was observed, splitting by any requested variables.
}
\details{
Averages for multiple statistics can be calculated at once. If the statistics
argument is set to c("pairwise", "single"), all calculated stats will be run.
If it is set to "single", then all non-pairwise statistics (pi, ho, maf, ect)
will be bootstrapped, if it is set to "pairwise", then all pairwise statistics
(fst) will be bootstrapped. Individual statistics currently cannot be
requested by name, since the computational differences to add additional types
is minimal.

The data can be broken up categorically by snp or sample metadata, as
described in \code{\link{Facets_in_snpR}}. Windows will only be calculated
using only SNPs on the same level of any provided facets. NULL and "all"
facets work as normally described in \code{\link{Facets_in_snpR}}.

As described in Hohelohe et al. (2010), the contribution of individual SNPs to
window averages can be weighted by the number of observations per SNP by
setting the nk argument to TRUE, as is the default. For bootstraps, nk values
are randomly drawn for each SNP in each window.

Centers for windows can either every SNP (if no step size is provided), or
every step kilobases from the 0 position of each snp level facet category
(chromosome, etc.).

The size of sliding windows are defined by the "sigma" argument. Note that
this value, as well as that provided to the "step" arguement, are given in
kilobases. Each window will include SNPs within 3*sigma kilobases from the
window center. Past this point, the effect of each additional SNP on the
window average would be very small, and so they are dropped for computational
efficiency (see Hohenlohe (2010)).
}
\examples{
\dontrun{
# add a few statistics
x <- calc_pi(stickSNPs, "chr.pop")
x <- calc_ho(x, "chr.pop")
x <- calc_pairwise_fst(x, "chr.pop")
# smooth with a fixed slide between window centers.
x <- calc_smoothed_averages(x, "chr.pop", sigma = 200, step = 50)
get.snpR.stats(x, "chr.pop", "single.window") # pi, ho
get.snpR.stats(x, "chr.pop", "pairwise.window") # fst
}
}
\references{
Hohenlohe et al. (2010). \emph{PLOS Genetics}
}
\author{
William Hemstrom
}
