% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_tajimas_d}
\alias{calc_tajimas_d}
\title{Tajima's D from SNP data.}
\usage{
calc_tajimas_d(x, facets, sigma, step, par = F)
}
\arguments{
\item{x}{Input data, in allele count format as given by format_snps output option 1.}

\item{step}{Lenght to slide between each window, in kb.}

\item{ws}{The size of each window, in kb.}

\item{report}{When reporting progress, how many windows should be calculated between each report?}
}
\value{
Data frame containing metadata, Waterson's Theta, Tajima's Theta, and Tajima's D for each window.
}
\description{
\code{Tajimas_D} calculates Tajima's theta/pi, Waterson's theta, and Tajima's D over a sliding window. Pi calculated as in Hohenlohe et al. 2010. Tajima's D is calculated using the formula from Tajima (1989) Statistical Method for Testing the Neutral Mutation Hypothesis by DNA Polymorphism.
}
\details{
Description of x:
   Must contain colums containing the number of *unique* alleles, total count of alleles sequenced in all individuals, and subsequent alleles counts for each observed allele in columns named "n_alleles", "n_total", "ni1", and "ni2". Also needs a column containing the position of each SNP, in bp. This matches the allele count/bayescan format as given by format_snps option one.
}
\examples{
#convert example data into the correct format and run splitting by groups and pops.
pops <- table(substr(colnames(stickSNPs[,4:ncol(stickSNPs)]), 1, 3))
ac <- format_snps(stickSNPs, 3, pop = pops)
run_gp(ac, Tajimas_D, 200, 50)

}
