% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_weighted_stat}
\alias{calc_weighted_stat}
\title{Weighted stat averages/SEs.}
\usage{
calc_weighted_stat(x, stat, boots = 30)
}
\arguments{
\item{x}{Input statistic data.}

\item{stat}{Name of the statistic for which to calculate a weighted mean and se.}

\item{boots}{Number of bootstraps to perform to calcluate se. 30 is usually sufficient.}
}
\value{
A data frame with columns for the population ID, weighted mean, and weighted average.
}
\description{
\code{calc_weighted_stat} calculates a weighted average statistic and standard errors for a variable via the bootstrapping method described by Gatz and Smith (1995), The standard error of a weighted mean concentration—I. Bootstrapping vs other methods.
}
\details{
Description of x:
   Must contain colums containing the statistic of interest (such as pi from calc_pi), population ID, and the weights (such as the number of alleles sequenced, or the n_total column from in the allele count format) for each SNP. These columns must be named to match the stat argument, "pop", and either "n_total" or "nk", respectively.
}
\examples{
calc_weighted_stat(randPI, "pi")

}
