% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check.snpR.facet.request}
\alias{check.snpR.facet.request}
\title{Get details on and clean up facet arguments.}
\usage{
check.snpR.facet.request(
  x,
  facets,
  remove.type = "snp",
  return.type = FALSE,
  fill_with_base = TRUE
)
}
\arguments{
\item{x}{snpRdata object to compare to}

\item{facets}{character. facets to check.}

\item{remove.type}{character. "snp", "sample", "complex", "simple" or anything
else, typically "none". Type of facet to remove.}

\item{return.type}{logical, default FALSE. If true, returns both facets and
facet types ("snp", "sample", "complex", or ".base") as a length two list.}

\item{fill_with_base}{logical, default TRUE. If true, fills the returned facets
with .base if nothing is left after facets are removed. Otherwise returns null in this case.}
}
\description{
Given a snpRdata object, this function cleans up and determines the types of
requested facets. Internal. If requested, can clean facet requests to purge a
specific facet type (snp, sample, complex). Used in functions that run only on
a specific type of facet.
}
\details{
Facet designation of NULL or "all" follows the typical rules.

Facets designated as described in \code{\link{Facets_in_snpR}}.
}
\author{
William Hemstrom
}
