% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check_calced_stats}
\alias{check_calced_stats}
\title{Check if a given stat or vector of stats have been calculated any number of
facets.}
\usage{
check_calced_stats(x, facets, stats)
}
\arguments{
\item{x}{snpRdata object to check}

\item{facets}{character. See \code{\link{Facets_in_snpR}}}

\item{stats}{character. Names of stats to check.}
}
\value{
A named list with an entry for each facet containing a named logical
  vector indicating if the provided stats have been calculated yet.
}
\description{
Check if a given stat or vector of stats have been calculated any number of
facets.
}
\author{
William Hemstrom
}
