% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{convert_2_to_1_column}
\alias{convert_2_to_1_column}
\title{Tiny internal used when converting from a phased, transposed dataset (like
from an MS output)}
\usage{
convert_2_to_1_column(x)
}
\arguments{
\item{x}{input matrix}
}
\description{
Tiny internal used when converting from a phased, transposed dataset (like
from an MS output)
}
