% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{cross_validate_genomic_prediction}
\alias{cross_validate_genomic_prediction}
\title{Run a single cross-validation with BGLR.}
\usage{
cross_validate_genomic_prediction(
  x,
  response,
  iterations = 10000,
  burn_in = 1000,
  thin = 100,
  cross_percentage = 0.9,
  model = "BayesB",
  cross_samples = NULL,
  plot = TRUE,
  interpolate = "bernoulli"
)
}
\arguments{
\item{x}{snpRdata object}

\item{response}{character. Name of the column containing the response
variable of interest. Must match a column name in sample metadata.}

\item{iterations}{numeric. Number of iterations to run the MCMC chain for.}

\item{burn_in}{numeric. Number of burn in iterations to run prior to the MCMC
chain.}

\item{thin}{numeric. Number of iterations to discard between each recorded
data point.}

\item{cross_percentage}{numeric, default 0.9. The proportion of sample to use
to create the model. Must be greater than 0 and less than 1.}

\item{model}{character, default "BayesB". Prediction model to use, see
description for the ETA argument in \code{\link[BGLR]{BGLR}}.}

\item{cross_samples}{numeric, default NULL. Optional vector of sample indices
to use for cross-validation. If provided, the cross_percentage argument
will be ignored.}

\item{plot}{logical, default TRUE. If TRUE, will generate a ggplot of the
cross-validation results.}
}
\value{
A list containing: \itemize{ \item{model: } The results from
  \code{\link{run_genomic_prediction}}. \item{model.samples: } Indices of the
  samples used to construct the model. \item{cross.samples: } Indices of the
  samples used to cross-validate the model. \item{comparison: } A data.frame
  containing the observed and predicted phenotypes/Breeding Values for the
  cross-validation samples. \item{rsq: } The r^2 value for the observed and
  predicted phenotypes/Breeding Values for the cross-validation samples. }
}
\description{
Run genomic prediction given a single response variable (usually a phenotype)
using the \code{\link[BGLR]{BGLR}} function. Unlike other snpR functions,
this returns the resulting model direcly, so overwrite with caution. This
will leave a specified portion of the samples out when running the model,
then perform cross-validation.
}
\details{
This function is provided as a wrapper to plug snpRdata objects into the
\code{\link[BGLR]{BGLR}} function in order to easily run genomic prediction
on a simple model where a single, sample specific meta data variable is
provided as the response variable. To do so, this function formats the data
into a transposed "sn" format, as described in \code{\link{format_snps}}
using the bernoulli method to interpolate missing genotypes. Several
different prediction models are available, see the documentation the ETA
argument in \code{\link[BGLR]{BGLR}} for details. Defaults to the "BayesB"
model, which assumes a "spike-slab" prior for allele effects on phenotype
where most markers have a very small effect size and a few can have a much
larger effect.

Unlike most snpR functions, this function does not support facets, since each
run can be very slow. Instead, an individual facet and facet level of
interest should be selected with \code{\link{subset_snpR_data}}. See
examples.

The portion of samples left out for cross-validation is defined by the
cross_percentage argument. Specifically, the proportion given will be used to
create the model, the remaining portion will be left out. For example, if
cross_percentage = 0.9, 90% of the samples will be used to run the model and
the remaining 10% will be used for cross-validation. Alternatively, if
cross_samples are provided, the specifed samples (by column index) will be
used used for cross validation instead. This may be useful for a systematic
leave-one-out cross-validation.


See documentation for \code{\link[BGLR]{BGLR}} for more details and for a
full list of references.
}
\examples{
# run and plot a basic prediction
## add some dummy phenotypic data.
dat <- stickSNPs
sample.meta(dat) <- cbind(weight = rnorm(ncol(stickSNPs)), sample.meta(stickSNPs))
## run cross_validation
cross_validate_genomic_prediction(dat, response = "weight", iterations = 1000, burn_in = 100, thin = 10)

}
\references{
Pérez, P., and de los Campos, G. (2014). \emph{Genetics}.
}
\author{
William Hemstrom
}
