% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping_functions.R
\name{do_bootstraps}
\alias{do_bootstraps}
\title{Generate bootstrapped windowed statistics}
\usage{
do_bootstraps(x, facets = NULL, boots, sigma, step = NULL,
  statistics = "all", nk = T, par = FALSE, do.p = TRUE,
  p.alt = "two-sided")
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character or NULL, default NULL. Categories by which to break up
bootstraps.}

\item{boots}{numeric. Number of bootstraps to generate.}

\item{sigma}{numeric. Designates the width of windows in kilobases. Full
window size is 6*sigma.}

\item{step}{numeric or NULL, default NULL. Designates the number of kilobases
between each window centroid. If NULL, windows are centered on each SNP.}

\item{nk}{logical, default TRUE. If TRUE, weights SNP contribution to window
averages by the number of observations at those SNPs.}

\item{par}{numeric or FALSE, default FALSE. If numeric, the number of cores to
use for parallel processing.}

\item{do.p}{logical, default TRUE. Determines if p-values should be calculated
for sliding windows.}

\item{p.alt}{character, default "two-sided". Specifies the alternative
hypothesis to be used. Options: \itemize{ \item "less": probability that a
bootstrapped value is as small or smaller than observed. \item "greater":
probability that a bootstrapped value is as large or larger than observed.
\item "two-sided": probability that a bootstrapped value is as or more
extreme than observed. }}

\item{statistic}{character. Designates the statistic(s) to smooth, typically
"all", "single", or "pairwise". See details for options.}
}
\value{
A snpRdata object with bootstrapped windows merged in to the
 window.bootstraps slot. If do.p is TRUE, it will also merge p values in for
 bootstrapped statistics into the stats or window.stats sockets.
}
\description{
\code{do_bootstraps} creates a distribution of bootstrapped smoothed values
for requested statistcs.
}
\details{
Bootstrapps are conducted as described by Hohenlohe et al. (2010). For each
bootstrap, this function draws random window position, then draws random
statistics from all provided SNPs to fill each observed position on that
window and calculates a smoothed statistic for that window using
gaussian-smoothing. Note that a "position" column \emph{must} be present in
the snp metadata of the snpRdata object to do any window calculations.

Bootstraps for multiple statistics can be calculated at once. If the
statistics argument is set to "all", all calculated stats will be run. If it
is set to "single", then all non-pairwise statistics will be bootstrapped, if
it is set to "pairwise", then all pairwise statistics will be bootstrapped.
Inidividual statistics can also be requested by name ("pi", "ho", etc.). All
statistics bootstrapped at the same time will be calculated using \emph{the
same randomly filled windows}, and thus do not represent independant
observations between statistics. This is still computationally more efficient,
however.

The data can be broken up categorically by snp or sample metadata, as
described in \code{\link{Facets_in_snpR}}.

As described in Hohelohe et al. (2010), the contribution of individual SNPs to
window averages can be weighted by the number of observations per SNP by
setting the nk argument to TRUE, as is the default. For bootstraps, nk values
are randomly drawn for each SNP in each window.

Possible centers for windows can either SNPs (if no step size is provided), or
every step kilobases from the 0 position of each snp level facet category
(chromosome, etc.).

If do.p is TRUE, calculates p-values for smoothed values of a statistic based upon the
bootstrapped null distribution of that statistic using an emperical continuous
distribution function.
}
\examples{
# add statistics
dat <- calc_basic_snp_stats(dat3, c("group.pop"), sigma = 200, step = 150)
do_bootstraps(dat, facets = c("group.pop"), boots = 1000, sigma = 200, step = 150)

}
\references{
Hohenlohe et al. (2010). \emph{PLOS Genetics}
}
\author{
William Hemstrom
}
