% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.fix..call}
\alias{.fix..call}
\title{Fixes calling scope warning in .. calls with data.table}
\usage{
.fix..call(fun)
}
\arguments{
\item{fun}{function to run}
}
\description{
Needed since ..x vars need to be defined as global variables for CRAN, which data.table will throw a warning about.
}
