% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata.R
\name{.get.snpR.stats}
\alias{.get.snpR.stats}
\title{Internal version of get.snpR.stats, for backwards comparability if type is set or
for fetching data for get.snpR.stats.}
\usage{
.get.snpR.stats(x, facets = NULL, type = "single", col_pattern = NULL)
}
\description{
See documentation for get.snpR.stats.
}
