% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.paste.by.facet}
\alias{.paste.by.facet}
\title{Paste together metadata according to a list of column names}
\usage{
.paste.by.facet(df, facets, sep = ".")
}
\arguments{
\item{df}{data.frame with data do paste}

\item{facets}{facets to paste together. Often produced by \code{\link{.split.facet}}. Can also be a numeric vector of columns to use.}

\item{sp}{character, default ".". Pasted facets will be split by this.}
}
\description{
Paste together metadata according to a list of column names
}
