% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{.split.facet}
\alias{.split.facet}
\title{Split a single or multiple compound facet into parts}
\usage{
.split.facet(facet)
}
\arguments{
\item{facet}{compound facet(s) to split}
}
\description{
Split a single or multiple compound facet into parts
}
