% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{.subset_snpR_data}
\alias{.subset_snpR_data}
\title{old subset version with different facet specification. For internal use.}
\usage{
.subset_snpR_data(
  x,
  snps = 1:nrow(x),
  samps = 1:ncol(x),
  facets = NULL,
  subfacets = NULL,
  snp.facets = NULL,
  snp.subfacets = NULL
)
}
\description{
old subset version with different facet specification. For internal use.
}
