% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{estimate_selection}
\alias{estimate_selection}
\title{Selection coefficients.}
\usage{
estimate_selection(p_before, p_after)
}
\arguments{
\item{p_before}{Intial frequency of allele \emph{p}.}

\item{p_after}{Frequency of allele \emph{p} after one generation.}
}
\value{
A numeric value, the estimate of \emph{s}.
}
\description{
Estimates the selection coefficient \emph{S} across one generation given intial and final allele frequencies, using an equation adapted from Gillespie (2010) Population genetics: a concise guide, equation 3.2.
}
\examples{
estimate_selection(.01, .011)

}
