% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata.R
\name{extract_snpRdata}
\alias{extract_snpRdata}
\alias{genotypes}
\alias{genotypes<-}
\alias{snp.meta}
\alias{snp.meta<-}
\alias{sample.meta}
\alias{sample.meta<-}
\title{Get from or overwrite components of a snpRdata object}
\usage{
genotypes(x)

genotypes(x) <- value

snp.meta(x, value)

snp.meta(x) <- value

sample.meta(x, value)

sample.meta(x) <- value
}
\arguments{
\item{x}{snpRdata object to get genotype data from.}

\item{value}{Genotypes, snp metadata, or sample metadata}
}
\description{
Fetch or overwirte the major parts of a snpRdata object (genotypes, snp meta,
or sample meta). If overwritten, any calculated stats will be removed, since
their values may be dependant upon changes in metadata.
}
\section{Functions}{
\itemize{
\item \code{genotypes}: view genotypes

\item \code{genotypes<-}: set genotypes

\item \code{snp.meta}: view snp meta

\item \code{snp.meta<-}: set snp meta

\item \code{sample.meta}: view sample meta

\item \code{sample.meta<-}: set sample meta
}}

\examples{
# copy test data
test <- stickSNPs

# show genotypes
genotypes(test)

# show or overwrite snp meta
snp.meta(test)
snp.meta(test) <- data.frame(pos = sample(10000, nrow(test), T), chr = sample(LETTERS[1:4], nrow(test), T))

#show or overwrite sample meta
sample.meta(test)
sample.meta(test) <- data.frame(fam = sample(LETTERS[1:4], ncol(test), T), pop = sample(LETTERS[5:8], ncol(test), T))
}
