% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{fetch.sample.meta.matching.task.list}
\alias{fetch.sample.meta.matching.task.list}
\title{Internal to find which sample metadata rows match a row from a task list.}
\usage{
fetch.sample.meta.matching.task.list(x, task.list.row)
}
\description{
Internal to find which sample metadata rows match a row from a task list.
}
