% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{fetch.snp.meta.matching.task.list}
\alias{fetch.snp.meta.matching.task.list}
\title{Internal to find which snp metadata rows match a row from a task list.}
\usage{
fetch.snp.meta.matching.task.list(x, task.list.row)
}
\arguments{
\item{x}{snpRdata object to check}

\item{task.list.row}{A row of an object created by
\code{\link{get.task.list}} to look up info about.}
}
\description{
Internal to find which snp metadata rows match a row from a task list.
}
