% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{filters}
\alias{filters}
\title{Report on filters used on a snpRdata object.}
\usage{
filters(x)
}
\arguments{
\item{x}{snpRdata object}
}
\value{
Prints easily human readable results to the console and also returns
a more machine readable string.
}
\description{
snpR will automatically track the methods used on a
snpRdata object. Using \code{\link{filters}} on that object will provide
details on the the filters used.
}
\details{
Printed outputs contain the filters used in the order they were applied
alongside the filtering stringency and any facets applied over, if 
applicable. Note that some output formats from \code{format_snps}, like the
\code{vcf} format, will notes on the filters used as well.
}
\examples{
# filter the data
x <- filter_snps(stickSNPs, 
                 min_ind = 0.75, 
                 min_loci = 0.75,
                 maf = 0.1,
                 maf_facets = "pop")

# fetch the filters used
filters(x)

}
\author{
William Hemstrom
}
