% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{find.snpR.facets}
\alias{find.snpR.facets}
\title{Grab information of the facets present in snpRdata objects.}
\usage{
find.snpR.facets(x)
}
\arguments{
\item{x}{snpRdata object.}
}
\description{
Internal, grabs via reference to face.meta socket rather than the facets
socket.
}
