% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{format_and_check_numeric}
\alias{format_and_check_numeric}
\title{Format numeric data in the 0000 or 000000 form.}
\usage{
format_and_check_numeric(gt, mDat)
}
\description{
Formats numeric data into NN format, checking for 3+ alleles at an loci.
Alleles coded as A and C. Expects individuals in columns.
}
