% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{Full_LD_w_groups}
\alias{Full_LD_w_groups}
\title{Pairwise LD from SNP data, split by group.}
\usage{
Full_LD_w_groups(x, ecs, prox_table = T, matrix_out = T, mDat = "NN",
  report = 1, sr = TRUE, chr.length = NULL, input = "SNP",
  ss = FALSE)
}
\arguments{
\item{x}{Input data, in either numeric or character formats, as given by format_snps output options 2 or 6.}

\item{ecs}{Number of extra metadata columns at the start of x.}

\item{prox_table}{If TRUE, a proximity table is produced.}

\item{matrix_out}{If TRUE, pairwise LD matrices are produced.}

\item{mDat}{Character variable matching the coding for missing *genotypes* in x (typically "N" or "00").}

\item{report}{Reports progress every report linkage groups/chromosomes.}

\item{sr}{boolean, default TRUE. Should within group progress reports be surpressed?}
}
\value{
Matrices containing rsq, Dprime, and p values for each SNP vs every other SNP. Can also produce a proximity table, which contains the rsq, Dprime, and p value for each pairwise comparison and the distance between the SNPs in those comparisons. Returns matrices and prox table as sequential elements in a named list. Prox table is concatenated across all groups, matrices are named first by group, second by LD measure.
}
\description{
\code{Full_LD_w_groups} calls \code{\link{LD_full_pairwise}} after spliting the data by a column named "group". Knits the results together, unlike if run with \code{\link{run_g}}.
}
\details{
Matrix outputs are properly formated for LD_pairwise_heatmap. See notes for \code{\link{LD_full_pairwise}}.

Description of x:
   Contains metadata in columns 1:ecs. Remainder of columns contain genotype calls for each individual. Each row is a different SNP, as given by format_snps output options 4 or 6. Requires the column containing the position of the loci in base pairs be named "position". Requires that the column containing linkage group/chromosome ID be named "group". Note that this *ignores populations*, split data into distinct populations before running.
}
\examples{
#returns prox table and LD matrices.
Full_LD_w_groups(stickSNPs[stickSNPs$group == "groupI" | stickSNPs$group == "groupII",1:53], ecs = 3)

}
