% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{fwe_correction}
\alias{fwe_correction}
\title{Correct for multiple testing.}
\usage{
fwe_correction(
  p,
  pcol = NULL,
  levs = c("subfacet", "facet"),
  methods = c("bonferroni", "holm", "BH", "BY"),
  case = c("overall", "by_facet", "by_subfacet")
)
}
\arguments{
\item{p}{data.frame/data.table containing p values and any facets to split by}

\item{pcol}{name of the column containing p values}

\item{levs}{names of the columns containing facet info to split by}

\item{methods}{methods to use}

\item{case}{case (overall, within each subfacet, or within each facet) at
which to apply corrections}
}
\description{
Correct for multiple testing.
}
