% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GO_functions.R
\name{gene_ave_stat}
\alias{gene_ave_stat}
\title{Interpolates stats at genes.}
\usage{
gene_ave_stat(x, y, stat)
}
\arguments{
\item{x}{Input gene data, with columns named "start", "end" and "probeID".}

\item{y}{Input stat data, typically from smoothed windows or (not recommended) raw statistics from SNPs.}

\item{stat}{Name of the statistic to interpolate.}
}
\description{
\code{gene_ave_stat} takes an input file of genes with start/end position noted and interpolates back the average statitstic of choice for that gene. Requires the "zoo" package.
}
\details{
Description of gene_data:
   Requires columns titled "group", "start", "end", and "probeID", containing gene linkage group/chromosome, start and end positions, and name, respectively.

Description of stat_data:
   Requires columns titled "group", "position", and one matching the stat argument, containing the linkage group/chromosome, position, and the value of the observed statistic at that position. Typically produced via windowed gaussian smoothing (from smoothed_ave).

Uses spline interpolation from the "zoo" package.
}
\examples{
gene_ave_stat(stickleGO, randSMOOTHed[randSMOOTHed$pop == "A",], "smoothed_pi")

}
