% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{genind.to.snpRdata}
\alias{genind.to.snpRdata}
\title{Convert genind object to snpRdata.}
\usage{
genind.to.snpRdata(genind, snp.meta = NULL, sample.meta = NULL)
}
\arguments{
\item{genind}{genind object.}

\item{snp.meta}{data.frame, default NULL. Metadata for each SNP, IDs from
genind may be added.}

\item{sample.meta}{data.frame, default NULL. Metadata for each individual
sample, IDs and pops from genind may be added.}
}
\description{
Convert genind object to snpRdata. Internal, called by import.snpR.data when
provided a adegenet genind object.
}
\author{
William Hemstrom
}
