% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{genlight.to.snpRdata}
\alias{genlight.to.snpRdata}
\title{Convert genlight object to snpRdata.}
\usage{
genlight.to.snpRdata(genlight, snp.meta = NULL, sample.meta = NULL)
}
\arguments{
\item{genlight}{genlight object.}

\item{snp.meta}{data.frame, default NULL. Metadata for each SNP, IDs from
genind may be added.}

\item{sample.meta}{data.frame, default NULL. Metadata for each individual
sample, IDs and pops from genind may be added.}
}
\description{
Convert genlight object to snpRdata. Internal, called by import.snpR.data
when provided a adegenet genlight object.
}
\author{
William Hemstrom
}
