% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata.R
\name{genotypes}
\alias{genotypes}
\title{Get genotypes from a snpRdata object.}
\usage{
genotypes(x)
}
\arguments{
\item{x}{snpRdata object to get genotype data from.}
}
\description{
Genotypes will be formatted with SNPs in rows and individuals in columns.
}
