% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata.R
\name{get.snpR.stats}
\alias{get.snpR.stats}
\title{Pull calculated statistics from snpRdata objects.}
\usage{
get.snpR.stats(x, facets = NULL, stats = "single", bootstraps = FALSE)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character or NULL, default NULL. Facets for which to fetch
data.}

\item{stats}{character or NULL, default NULL. Statistics for which to fetch
data. Named identically as in the function used to calculate them (such as
calc_pi). See description for list. Alternatively, a type of statistic can
be requested--see description for details.}

\item{bootstraps}{logical, default FALSE. If FALSE, bootstraps will not be
returned even if they have been created for the requested statistics. Since
bootstrapped datasets are often quite large, returning these may use a lot
of additional memory.}
}
\description{
A convenience function that pulls statistics of any specified type at
particular facets from a snpRdata object.
}
\details{
Facets are specified as described in \code{\link{Facets_in_snpR}}. If facets
= "all", data for all facets, including the base facet, will be returned. By
default, the base facet alone will be returned.

Different statistics are returned either by named statistic or by type.
}
\section{Types}{

\itemize{
\item single: non-pairwise, non-window statistics (pi, ho,
etc.)
\item pairwise: pairwise, non-window statistics (Fst).
\item single.window: non-pairwise, sliding window statistics.
\item pairwise.window: pairwise, sliding window statistics.
\item LD: linkage disequilibrium matrices and tables.
\item bootstraps: bootstraps
of window statistics.
\item genetic_distance: genetic distances
\item allele_frequency_matrix: allele frequency matrices.
\item geo_dist: geographic distances.
\item ibd: isolation by distance results.
\item sample: sample stats.
\item pop: population summary statistics.
\item weighted.means: Weighted means of SNP statistics per facet level.
}
}

\section{Statistics}{

\itemize{
\item ho: Observed heterozygosity, via \code{\link{calc_ho}}.
\item pi: pi (expected number of pairwise
differences, via \code{\link{calc_pi}}).
\item maf: minor allele
frequencies (and major/minor counts and identities), via
\code{\link{calc_maf}}.
\item private: private allele identities, via
\code{\link{calc_private}}.
\item association: results from phenotypic
association tests, via \code{\link{calc_association}}.
\item hwe: Hardy-Weinberg Equilibrium p-values, via \code{\link{calc_hwe}}.
\item tajimas_d: Tajima's D, Watterson's Theta, and Tajima's Theta, via
\code{\link{calc_tajimas_d}}.
\item fst: Pairwise Fst, via
\code{\link{calc_pairwise_fst}}.
\item het_hom_ratio: Heterozygote/Homozygote ratios within individuals, via
\code{\link{calc_het_hom_ratio}}.
\item ne: Effective population size
estimates, via \code{\link{calc_ne}}.
\item ld: Pairwise linkage
disequilibrium, via \code{\link{calc_pairwise_ld}}.
\item genetic_distances: Genetic distances,
\code{\link{calc_genetic_distances}}.
\item isolation_by_distance: Genetic isolation by distance metrics, via
\code{\link{calc_isolation_by_distance}}.
\item geographic_distance: Geographic distances between samples--does not use
genetic data. Calculated
during \code{\link{calc_isolation_by_distance}}, but fetchable
independently here.
\item random_forest: Random forest snp-specific responses, see
\code{\link{run_random_forest}}.
}
}

\examples{
# generate some statistics
dat <- calc_pi(stickSNPs, "pop")
dat <- calc_pairwise_fst(dat, "pop")

# fetch pi
get.snpR.stats(dat, "pop", "pi")
# fetch fst
get.snpR.stats(dat, "pop", "fst")
# fetch both
get.snpR.stats(dat, "pop", c("pi", "fst"))

# return a type of statistic instead of specific statistics
dat <- calc_ho(stickSNPs, "chr.pop")
get.snpR.stats(dat, "chr.pop", "weighted.means")

}
\author{
William Hemstrom
}
