% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_GWAS.R
\name{get_GWAS}
\alias{get_GWAS}
\title{Download and parse GWAS catalog data}
\usage{
get_GWAS(catalog = "download", filter.criteria = NA)
}
\arguments{
\item{catalog}{character; either a file path to a previously downloaded
version of the GWAS catalog or "download", to download the latest version.
[Default = "download"]}

\item{filter.criteria}{character; filtering criteria - list of terms, separated
bu '|' as a single character string, e.g. "diabetes | ischaemic | stroke". If no
filter is chosen, then all SNPs are outputted [Default = NA]}
}
\value{
a data.frame of the gwas catalog pruned
}
\description{
This function will download the latest version of the GWAS
catalog from UCSC with hg38 co-ordinates, retain only those
SNPs with a genome-wide association p-value less than 5x10-8, and then
filter this further to create a data.frame containing SNPs determined
by the user-specified filtering criteria.
}
