% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{HWE}
\alias{HWE}
\title{Hardy-Weinburg Equilibrium from SNP data.}
\usage{
HWE(x, ecs, miss = "NN", test = "exact", n.reps = 20000)
}
\arguments{
\item{x}{Input data, in either NN or 0000 format, as given by format_snps output options 4 or 6.}

\item{ecs}{The number of extra columns at the start of the input data.frame containing metadata.}

\item{miss}{Characters which code for missing *genotypes* ("NN" or "0000", for example).}

\item{test}{Which test to use? "exact" for the exact test, "permutation" for permutation test.}

\item{n.reps}{Number of reps to use for permutation test.}
}
\value{
A numeric vector containing p-values, which are the probability that a SNP in HWE would be as or more diverged from the expected equilibrium genotype frequencies. In the same order as input.
}
\description{
Function to test for deviation from HWE for each SNP via either permutation or the exact test given by Wigginton et al 2005.
}
\details{
Part of this code is edited from Wigginton, JE, Cutler, DJ, and Abecasis, GR (2005) A Note on Exact Tests of Hardy-Weinberg Equilibrium. American Journal of Human Genetics. 76: 000 - 000. code available at http://csg.sph.umich.edu/abecasis/Exact/snp_hwe.r

Description of x:
   Contains metadata in columns 1:ecs. Remainder of columns contain genotype calls for each individual. Each row is a different SNP, as given by format_snps output options 4 or 6. Note that this *ignores populations*, split data into distinct populations before running.
}
\examples{
HWE(stickSNPs, 3)
}
