% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_plink.R
\name{install_plink}
\alias{install_plink}
\title{Function to download PLINK binary}
\usage{
install_plink()
}
\value{
Path to PLINK binary
}
\description{
Function to download PLINK binary
}
\examples{
plink <- install_plink()
}
