% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{interpolate_sn}
\alias{interpolate_sn}
\title{Interpolate sn formatted data.}
\usage{
interpolate_sn(sn, method = "bernoulli")
}
\arguments{
\item{sn}{data.frame. Input sn formatted data, as produced by
\code{\link{format_snps}}. Note that \code{\link{format_snps}} has an option
to automatically call this function during formatting.}

\item{method}{character, default "bernoulli". Method to used for
interpolation, either bernoulli or af. See details.}
}
\description{
An internal function to interpolate sn formatted data using either the
bernoulli or expected minor allele count approaches. Typically entirely
internal, called via format_snps.
}
\details{
Interpolating missing data in sn formatted data is useful for PCA, genomic
prediction, tSNE, and other methods. Specify method = "af" to insert the
expected number of minor alleles given SNP allele frequency or "bernoulli" to
do binomial draws to determine the number of minor alleles at each missing
data point, where the probability of drawing a minor allele is equal to the
minor allele frequency. The expected number of minor alleles based on the
later method is equal to the interpolated value from the former, but the later
allows for multiple runs to determine the impact of stochastic draws and is
generally prefered and required for some downstream analysis. It is therefore
the default.
}
\author{
William Hemstrom
}
