% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{is.snpRdata}
\alias{is.snpRdata}
\title{Check if an object is a snpRdata object.}
\usage{
is.snpRdata(x)
}
\arguments{
\item{x}{object to check}
}
\value{
Logical, TRUE if the object is a snpRdata object.
}
\description{
Checks if an object is a snpRdata object.
}
\author{
William Hemstrom
}
