% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{LD_full_pairwise}
\alias{LD_full_pairwise}
\title{Pairwise LD from SNP data.}
\usage{
LD_full_pairwise(x, ecs, prox_table = TRUE, matrix_out = TRUE,
  mDat = "NN", sr = FALSE, input = "SNP", chr.length = NULL,
  levels = FALSE, par = FALSE, ss = FALSE, level_report = 1,
  maf = FALSE)
}
\arguments{
\item{x}{Input data, usually in either numeric or character formats for SNP data, as given by format_snps output options "numeric" or "character". ms filepath or phased haplotypes (such as from an imported ms output) also accepted.}

\item{ecs}{Number of extra metadata columns at the start of x.}

\item{prox_table}{If TRUE, a proximity table is produced.}

\item{matrix_out}{If TRUE, pairwise LD matrices are produced.}

\item{mDat}{Character variable matching the coding for missing *genotypes* in x (typically "NN" or "0000").}

\item{sr}{boolean, default FALSE. Should progress reports be surpessed?}

\item{input}{Character, default "SNP". Input data type. Options: "SNP", as given by format_snps output options "numeric" or "character". "ms": filepath to ms formatted data. "haplotype": imported ms data, where each column is a fully phased gene copy/chromosome.}

\item{chr.length}{Numeric, default NULL. Length of chromosomes, for ms inputs.}

\item{levels}{Levels to split the data by for LD calculations, such as chromosome or linkage group. Must match in input column name.}

\item{par}{Numeric or FALSE, default FALSE. Number of cores to parallelize the analysis by.}

\item{ss}{Numeric or FALSE, default FALSE. Number or proportion of SNPs to subsample for LD analysis.}

\item{level_report}{Numeric, default 1. Progress through levels is reported every level_report levels.}

\item{maf}{Numeric or FALSE, default FALSE. Minor allele frequency filter cuttoff, only for ms or haplotype data.}
}
\value{
Matrices containing rsq, Dprime, and p values for each SNP vs every other SNP. Can also produce a proximity table, which contains the rsq, Dprime, and p value for each pairwise comparison and the distance between the SNPs in those comparisons. Returns matrices and prox table as sequential elements in a named list.
}
\description{
\code{LD_full_pairwise} calculates LD between each pair of SNPs. If called as is, assumes one linkage group/chromosome and one population!
}
\details{
Matrix outputs are properly formated for LD_pairwise_heatmap.

Description of x:
   Contains metadata in columns 1:ecs. Remainder of columns contain genotype calls for each individual. Each row is a different SNP, as given by format_snps output options 4 or 6. Requires the column containing the position of the loci in base pairs be named "position". Note that this \emph{ignores populations}, split data into distinct populations before running.
}
\examples{
#returns prox table and LD matrices.
LD_full_pairwise(stickSNPs[stickSNPs$group == "groupI",1:53], ecs = 3)

}
