% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{LD_pairwise_heatmap}
\alias{LD_pairwise_heatmap}
\title{Create a heatmap from pairwise linkage data.}
\usage{
LD_pairwise_heatmap(x, r = NULL, l.text = "rsq", colors = c("white",
  "black"), title = NULL, t.sizes = c(16, 13, 10, 12, 10),
  background = "white")
}
\arguments{
\item{x}{Data, format like that given by LD_full_pairwise rsq and Dprime outputs. Column names must be snp positions, as must either the first column or row names. Alternatively, a named list of the several such objects to plot and their titles can be provided.}

\item{r}{Region of the chromosome to subset and plot. Given in kb in the format numeric vector c(lower, upper).}

\item{l.text}{Legend title.}

\item{colors}{Colors to use in tiles, character vector format c("lower", "upper").}

\item{title}{Plot title.}

\item{t.sizes}{Text sizes, numeric vector format c(title, legend.title, legend.ticks, axis, axis.ticks)}
}
\value{
A pairwise LD heatmap as a ggplot object. If multiple objects were provided to plot, these will be included as objects.
}
\description{
\code{LD_pairwise_heatmap} prepares a ggplot2 heatmap from pairwise LD SNP data, in the format of that given by \code{\link{LD_full_pairwise}}. Should also work on pairwise FST data. Darker values are higher. Special thanks to Nicholas Sard for part of this code!
}
\details{
Since the output is a ggplot object, options can be added or changed by adding "+ function()" to the end of \code{LD_pairwise_heatmap}. Some common options are also built into this function as agruments, but can be overwritten freely.

Output from \code{\link{LD_full_pairwise}} runs directly or after saving and re-import without issue. Position data for the rows can be saved either as rownames or as the first column.

This function was partially written by Nicholas Sard.
}
\examples{
#base plot
LD_pairwise_heatmap(stickLD, title = "ASP group IX Pairwise LD")

#Change colors with agruments, add lines at SNP 10 using ggplot functions.
LD_pairwise_heatmap(stickLD, colors = c("green", "red"), title = "ASP group IX Pairwise LD") + geom_vline(xintercept = 10, color = "blue") + geom_hline(yintercept = 10, color = "blue")

}
