% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_SNPs.R
\name{ld_SNPs}
\alias{ld_SNPs}
\title{Identify SNPs in LD}
\usage{
ld_SNPs(plink, genotypeData, independent_snps, r2 = 0.8)
}
\arguments{
\item{plink}{Path to PLINK (v1.9)}

\item{genotypeData}{Path to the directory containing the PLINK binary genotype formatted 1000 
Genomes Project data which may be downloaded as a tar.gz file from this link: 
\url{http://www.broadinstitute.org/mpg/depict/depict_download/1kg/1000_genomes_project_phase3_CEU.tar.gz}.}

\item{independent_snps}{data.table object created by \code{prune_SNPs}}

\item{r2}{numeric value 0-1 indicating the r2 value to calculate SNPs in LD}
}
\value{
a data.table object
}
\description{
This function identifies all SNPs in LD with a list of independent_snps
generated using the \code{pruneSNPs()} function
}
