% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_SNPs2BED.R
\name{ld_SNPs2BED}
\alias{ld_SNPs2BED}
\title{Create BED file of SNP loci}
\usage{
ld_SNPs2BED(ld_snps, file = NA)
}
\arguments{
\item{ld_snps}{list object created by \code{ld_SNPs}}

\item{file}{character string specifying name of output BED file}
}
\value{
a data.frame object
}
\description{
Create BED file of SNP loci
}
