% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfs_functions.R
\name{make_SFS}
\alias{make_SFS}
\title{Generate a 1-2d site frequency spectrum from a dadi input file.}
\usage{
make_SFS(x, pops, projection, fold = F)
}
\arguments{
\item{x}{character or data.frame. Either a path to a dadi formatted input
file or a data.frame containing previously imported dadi formatted data.}

\item{pops}{character. A vector of population names of up to length 2
containing the names of populations for which the an SFS is to be created.}

\item{projection}{numeric. A vector of sample sizes to project the SFS to, in
\emph{number of gene copies}. Sizes too large will result in a SFS
containing few or no SNPs.}

\item{fold}{logical, default FALSE. Determines if the SFS should be folded or
left polarized.}
}
\value{
A matrix or vector containing the site frequency spectrum with a
  "pops" attribute containing population IDs, such as c("POP1", "POP2"). For
  a 2d SFS, the first pop is the matrix columns and the second is the matrix
  rows.
}
\description{
Generates a 1 or 2 dimensional site frequency spectrum from a dadi input file
using the projection methods and folding methods of Marth et al (2004) and
Gutenkunst et al (2009). This code is essentially an R re-implementation of
the SFS construction methods implemented in the program \emph{dadi} (see
Gutenkunst et al (2009)).
}
\details{
Site frequency spectrums are constructed using the projection methods
detailed in Marth et al (2004) and the 2 dimensional expansion in Gutenkunst
et al (2009). Folding methods are also taken from Gutenkunst et al (2009).
Either 1 or 2d SFSs can be constructed by providing a vector of population
names and projection sizes.
}
\references{
Gutenkunst et al (2009). Inferring the joint demographic history
  of multiple populations from multidimensional SNP frequency data.
  \emph{PLoS genetics}, 5(10), e1000695.

Marth et al (2004). The allele frequency spectrum in genome-wide
  human variation data reveals signals of differential demographic history in
  three large world populations. \emph{Genetics}, 166(1), 351-372.
}
