% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_snps.R
\name{matching_SNPs}
\alias{matching_SNPs}
\title{Create a set of matching SNP loci}
\usage{
matching_SNPs(snpdatabase = NULL, nperm = 1000, ld.snps)
}
\arguments{
\item{snpdatabase}{character string specifying path to the snp database collection.}

\item{nperm}{Number of permutations to perform (i.e. sets of matching SNPs to create)}

\item{ld.snps}{list object created by \code{ld_snps}}
}
\value{
List of GRanges whose length is equal to nperm
}
\description{
This function creates a set of matching SNP loci using data from 
the 1000 Genomes Phase 3 Project.  The SNPs in each set are matched for MAF, 
number of genes in the locus and LD 'buddies' at r2 0.8.
}
\details{
The snpdatabase is an rdata object containing all snps in the 1000G 
Phase3 along with the MAF, number of genes in the locus boundary 
(defined by r2 0.8) and number of ld buddies, split by MAF bin.  It may be
downloaded from \url{https://www.dropbox.com/s/s9roomzkq98zvhb/collection.rda?dl=0}.
}
