% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{merge.lists}
\alias{merge.lists}
\title{Merge lists element-to-element}
\usage{
\method{merge}{lists}(
  list1,
  list2,
  possible_end_level_names = c("Dprime", "rsq", "pval", "CLD", "S")
)
}
\arguments{
\item{list1}{list. The first list. Elements with identical names at all
levels in list 2 will be *replaced*.}

\item{list2}{list. The second list. Elements in list 2 with identical names
found in list 1 will replace those elements.}

\item{possible_end_level_names}{vector of possible stopping point names
(bottom level)}
}
\description{
Merges list 1 into list 2, adding any elements without matching equivalents
to list 2.
}
\details{
Merges element to element. Elements in list 1 with matching elements in list
2 will be replaced.

Do this by getting the names and "pathways" of all of the deepest level
objects, then looping through list 2 data and adding from list 1 that isn't
present at the same "pathway".
}
\author{
William Hemstrom
}
