% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{merge.snpR.stats}
\alias{merge.snpR.stats}
\title{Merge newly calculated stats into a snpRdata object.}
\usage{
\method{merge}{snpR.stats}(x, stats, type = "stats")
}
\arguments{
\item{x}{snpRdata object}

\item{stats}{data.frame/table/list. New data to be added to the existing
snpRdata object, x.}

\item{type}{character, default "stats". The type of statistic to merge, see
list in description.}
}
\description{
Internal function to quickly and accurately merge newly calculated statistics
into a snpRdata object. This should never be called externally. Takes and
returns the same snpRdata object, with new data.
}
\details{
Mostly relies on the smart.merge subfunction, which must be edited with care
and testing. LD merging is independant.

Type options: stats, pairwise, window.stats, pairwise.window.stats, and LD,
corresponding to slots of the snpRdata S4 object class.

For examples, see functions that use merge.snpR.stats, such as calc_pi or
calc_pairwise_ld.
}
