% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata.R
\name{nsamps}
\alias{nsamps}
\alias{ncol,snpRdata-method}
\title{Get the number of samples in a snpRdata object.}
\usage{
nsamps(x, value)

\S4method{ncol}{snpRdata}(x)
}
\arguments{
\item{x}{snpRdata object}
}
\description{
Get the number of samples in a snpRdata object.
}
\section{Methods (by class)}{
\itemize{
\item \code{snpRdata}: identical to ncol
}}

