% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata.R
\name{nsnps}
\alias{nsnps}
\alias{nrow,snpRdata-method}
\title{Get the number of snps in a snpRdata object.}
\usage{
nsnps(x, value)

\S4method{nrow}{snpRdata}(x)
}
\arguments{
\item{x}{snpRdata object}
}
\description{
Get the number of snps in a snpRdata object.
}
\section{Methods (by class)}{
\itemize{
\item \code{snpRdata}: identical to nrow
}}

