% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth.R
\name{optimum_slide_functions}
\alias{optimum_slide_functions}
\alias{optimum_scaled_slide}
\alias{optimum_fixed_slide}
\title{Increase the selection optimum each generation.}
\usage{
optimum_scaled_slide(opt, iv, slide = 0.3, ...)

optimum_fixed_slide(opt, slide = 0.3, ...)
}
\arguments{
\item{opt}{numeric. Initial optimum phenotype.}

\item{iv}{numeric. Initial variance. The optimum will increase by some
portion of this variance.}

\item{slide}{numeric. Proportion of the initial variance by which the optimum
phenotype will slide.}
}
\value{
The optimum phenotypic value in the next generation.
}
\description{
Increase selection optimum by either a given percentage of 
starting variance each generation or by a fixed amount. For use in
\code{\link{simulate_populations}}.
}
\details{
While two slide options are provided, others can be used instead by writing
functions which take the argument \code{opt}, the starting optimum in the
generation previous. Optionally, they
may also take the argument \code{iv}, which is the initial phenotypic
variation in the population, passed automatically by
\code{\link{simulate_populations}}. All functions must also take `...` to
allow for unused extra arguments pased by \code{\link{simulate_populations}},
such as \code{iv} for slides which do not use \code{iv}. Call
\code{optimum_scaled_slide} without parentheses to see an example.

Options:
\itemize{\item{optimum_scaled_slide: }{Change the optimum phenotype 
each generation by an amount that depends on the initial 
ammount of phenotypic variation in the population}
\item{optimum_fixed_slide: } {Change the optimum phenotype 
by a fixed amount each generation}}.
}
\section{Functions}{
\itemize{
\item \code{optimum_scaled_slide()}: slide scaled by genetic variance

\item \code{optimum_fixed_slide()}: fixed slide each generation

}}
