% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping_functions.R
\name{p_calc_boots}
\alias{p_calc_boots}
\title{Caculate p-values from bootstrapped distributions.}
\usage{
p_calc_boots(x, dist, alt = "two-sided")
}
\arguments{
\item{x}{Numeric vector of smoothed values for which to calculate p-values.}

\item{dist:}{Numeric vector of bootstrapped smoothed values.}

\item{alt:}{Character string. Which alternative hypothesis should be used? Options: #' \itemize{
   \item "less": probability that a bootstrapped value is as small or smaller than observed.
   \item "greater": probability that a bootstrapped value is as large or larger than observed.
   \item "two-sided": probability that a bootstrapped value is as or more extreme than observed.
}}
}
\value{
A numeric vector containing p-values for each input smoothed value.
}
\description{
\code{p_calc_boots} finds p-values for observed \emph{smoothed} statistics from bootstrapped distributions, such as that produced by \code{\link{resample_long}}.
}
\details{
Calculates p-values for smoothed values of a statistic based upon a bootstrapped null distribution of that statistic, such as those produced by resample_long using an emperical continuous distribution function.
}
\examples{
p_calc_boots(randSMOOTHed[randSMOOTHed$pop == "A",]$smoothed_pi, randPIBOOTS)
p_calc_boots(randSMOOTHed[randSMOOTHed$pop == "A",]$smoothed_pi, randPIBOOTS, alt = "less")

}
