% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{PCAfromPA}
\alias{PCAfromPA}
\title{PCAs from genetic data.}
\usage{
PCAfromPA(x, ecs, do.plot = "pop", c.dup = FALSE, mc = FALSE,
  counts = FALSE)
}
\arguments{
\item{x}{Input presence/absence data, as described in details.}

\item{ecs}{The number of extra metadata columns at the start of the input. Must be more that two to avoid errors. I really should fix that at some point. Includes the "pop" collumn.}

\item{do.plot}{FALSE or character vector, default "pop". If FALSE, no plot is produced. Up to two variables to be plotted with names corresponding to column names in x be given as a character vector.}

\item{c.dup}{boolean, default FALSE. Should duplicate individuals be searched for and removed? This is very slow if the data set is large!}

\item{mc}{If the data is filtered of poorly sequenced individuals, how many loci must the individuals have to be kept?}

\item{counts}{Numeric vector containing the number of loci sequenced per individual.}
}
\value{
A list containing the raw PCA output and a PCA plot in the form of a ggplot graphical object. The plot can be changed as usual with ggplot objects.
}
\description{
\code{PCAfromPA} creates a ggplot object PCA from presence/absense allelic data. If individuals which were sequenced at too few loci are to be filtered out, both the mc and counts argument must be provided.
}
\details{
Description of x:
   SNP or other allelic data in presence/absence format, as given by \code{\link{format_snps}} option 7. An additional column of population IDs titled "pop" must also be provided.
}
\examples{
PCAfromPA(stickPA, 2)
}
