% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_diagnostic}
\alias{plot_diagnostic}
\title{Basic diagnostic plots}
\usage{
plot_diagnostic(
  x,
  facet = NULL,
  projection = floor(nsnps(x)/1.2),
  fold_sfs = TRUE
)
}
\arguments{
\item{x}{snpRdata object}

\item{facet}{character, default NULL. Categorical metadata variables by which
to break up plots. Note that only one facet is allowed here. Missingness and
the PCA will have individuals colored by the given sample facet. See
\code{\link{Facets_in_snpR}} for more details.}

\item{fold_sfs}{logical, default TRUE. Determines if the SFS should be folded
or left polarized. If FALSE, snp metadata columns named "ref" and "anc"
containing the identity of the derived and ancestral alleles, respectively,
should be present for polarization to be meaningful.}

\item{project}{integer, default floor(nsnps(x)/1.2). A sample size to project
the SFS to, in \emph{number of gene copies}. Sizes too large will result in
a SFS containing few or no SNPs.}
}
\value{
A named list of diagnostic ggplot2 plots.
}
\description{
Create a suite of basic diagnostic plots (FIS density, 1D SFS, maf density,
PCA, and % missing data per individual) to describe the condition of the data
in a snpRdata object.
}
\examples{
# missingness and pca colored by pop
plot_diagnostic(stickSNPs, "pop")
}
\author{
William Hemstrom
}
