% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_pairwise_LD_heatmap}
\alias{plot_pairwise_LD_heatmap}
\title{Create a heatmap from pairwise linkage data.}
\usage{
plot_pairwise_LD_heatmap(
  x,
  facets = NULL,
  snp.subfacet = NULL,
  sample.subfacet = NULL,
  LD_measure = "CLD",
  r = NULL,
  l.text = "rsq",
  viridis.option = "inferno",
  title = NULL,
  t.sizes = c(16, 13, 10, 12, 10),
  background = "white"
)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character, default NULL. Categorical metadata variables by which
to break up plots. Must match facets for which LD data has been previously
calculated, and only a single facet can be plotted at once. See
\code{\link{Facets_in_snpR}} for more details.}

\item{snp.subfacet}{character, default NULL. Specific snp-specific levels of
the provided facet to plot. See examples.}

\item{sample.subfacet}{character, default NULL. Specific sample-specific
levels of the provided facet to plot. See examples.}

\item{LD_measure}{character, default rsq. LD metric to plot. Must be present
in the calculated LD data.}

\item{r}{Numeric. Region of the chromosome to subset and plot. Given in kb in
the format numeric vector c(lower, upper).}

\item{l.text}{character, default "rsq". Legend title.}

\item{viridis.option}{character, default "inferno". Viridis color scale option to use.
Other color scales may be subsituted by appending the scale_color_continuous
and scale_fill_continuous ggplot functions to the produced plot using the
'+' operator. See \code{\link[ggplot2]{scale_gradient}} for details.}

\item{title}{character. Plot title.}

\item{t.sizes}{numeric, default c(16, 13, 10, 12, 10). Text sizes, given as
c(title, legend.title, legend.ticks, axis, axis.ticks).}
}
\value{
A list containing: \itemize{ \item{plot: } A pairwise LD heatmap as a
 ggplot object. \item{dat: } Data used to generate the ggplot object. }
}
\description{
Prepares a ggplot2 heatmap from pairwise linkage disequilibrium data stored in
a snpRdata object.
}
\details{
Since the output is a ggplot object, options can be added or changed by adding
"+ function()" to the end of \code{LD_pairwise_heatmap}. Some common options
are also built into this function as agruments, but can be overwritten freely.

Specific facets and facet levels can be requested as long as LD data for those
facets has been previously calculated. Facets should be provided as described
in \code{\link{Facets_in_snpR}}. Only one facet can be plotted at once. Plots
for individual SNP or sample-specific facet levels for the specified facet can
also be requested. See examples. For facets with many categories, this is
strongly recommended, since plots with many facet levels become
computationally challenging to produce and difficult to interpret. A facets
argument of NULL will plot the base level facet, a facet argument of "all" is
not accepted.

Note that NA LD values will be colored white in the resulting plot.
}
\examples{
# get LD data
dat <- calc_pairwise_ld(stickSNPs, c("pop.group"))

# produce plots for linkage group IX in the ASP and CLF populations.
plot_pairwise_LD_heatmap(dat3, c("pop.group"), "groupIX", c("ASP", "CLF"))

# produce plots for every population for linkage group IV
plot_pairwise_LD_heatmap(dat3, c("pop.group"), "groupIV")


}
\author{
William Hemstrom

Nicholas Sard
}
