% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth.R
\name{population_growth_functions}
\alias{population_growth_functions}
\alias{logistic_growth}
\alias{BL_growth}
\title{Calculate population size changes each generation.}
\usage{
logistic_growth(n, K, r, ...)

BL_growth(n, B, ...)
}
\arguments{
\item{n}{numeric, starting population size.}

\item{K}{numeric, carrying capacity.}

\item{r}{numeric, logistic growth rate.}

\item{B}{numeric, the number of offspring each individual has}
}
\value{
The number of individuals in the next generation.
}
\description{
Increase population sizes each generation according to specific growth
rate equations. For use in \code{\link{simulate_populations}}.
}
\details{
Sevaral survival probablity distribution curves are provided. Each are coded
such that they expect \code{n}, the starting population size and end in
the `...` argument so that additional, unused arguments can be passed without
error. Other distributions can be coded in the same way if needed; call
\code{logistic_growth} without parentheses to see an example.

Options:
\itemize{\item{logistic_growth: }{Classical logistic growth model.}
\item{BL_growth: } {Growth based on Burger and Lynch (1995). Assumes
each individual has \code{B} offspring (so each pair has 2*B). 
Density dependence must be supplied elsewhere, such as to the 
\code{K_thin_post_surv} argument to 
\code{\link{simulate_populations}}
}}.
}
\section{Functions}{
\itemize{
\item \code{logistic_growth()}: logisitic growth

\item \code{BL_growth()}: Burger and Lynch (1995) growth

}}
