% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{process_FSTAT}
\alias{process_FSTAT}
\title{Convert a FSTAT file into a snpRdata object.}
\usage{
process_FSTAT(FSTAT_file, snp.meta = NULL, sample.meta = NULL, mDat = "0000")
}
\arguments{
\item{FSTAT_file}{character. A path to a genepop file.}

\item{snp.meta}{data.frame or null, default null. snp metadata.}

\item{sample.meta}{data.frame or null, default null. sample metadata}

\item{mDat}{character, default "0000". Missing data format. If the genotype
is 6 instead of 4 characters, will be automatically changed to 000000 IF
the default value is provided.}
}
\description{
Convert a FSTAT file into a snpRdata object.
}
\author{
William Hemstrom
}
