% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{process_ms}
\alias{process_ms}
\title{Internal to process a ms file}
\usage{
process_ms(x, chr.length)
}
\arguments{
\item{x}{filepath to ms file}

\item{chr.length}{length of the chromosome. If a single value, assumes all the same length.
If a vector of the same length as number of chr, assumes those are the chr lengths in order of apperance in ms file.}
}
\description{
Internal to process a ms file
}
\author{
William Hemstrom
}
