% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{process_plink}
\alias{process_plink}
\title{Read in plink data}
\usage{
process_plink(plink_file)
}
\arguments{
\item{plink_file}{extensionless filepath to plink files}
}
\description{
Read in plink data
}
\author{
William Hemstrom
}
