% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{process_vcf}
\alias{process_vcf}
\title{Convert a vcf file/vcfR object into a snpRdata object}
\usage{
process_vcf(vcf_file, snp.meta = NULL, sample.meta = NULL)
}
\arguments{
\item{vcf_file}{character or vcfR object. Either a path to a vcf file or a
vcfR object.}

\item{snp.meta}{data.frame or null, default null. snp metadata, will
overwrite data in vcf if provided.}

\item{sample.meta}{data.frame or null, default null. sample metadata, will
overwrite sample names in vcf if provided.}
}
\description{
Convert a vcf file/vcfR object into a snpRdata object
}
\author{
William Hemstrom
}
