% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_SNPs.R
\name{prune_SNPs}
\alias{prune_SNPs}
\title{Function to generate independent SNPs}
\usage{
prune_SNPs(plink = NULL, snps, genotypeData)
}
\arguments{
\item{plink}{Path to PLINK (v1.9)}

\item{snps}{character string to file containing output from \code{get_GWAS} or
the name of the data.table object created by \code{getGWAS}}

\item{genotypeData}{Path to the directory containing the PLINK binary genotype formatted 1000 
Genomes Project data which may be downloaded as a tar.gz file from this link: 
\url{http://www.broadinstitute.org/mpg/depict/depict_download/1kg/1000_genomes_project_phase3_CEU.tar.gz}.}
}
\value{
data.table object
}
\description{
Given a list of SNPs, this function will utilise PLINK to generate independent
SNPs (r2 < 0.2 and at least 500kb apart).  It utilises the
}
