% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping_functions.R
\name{resample_long}
\alias{resample_long}
\title{Bootstrapped smoothed values.}
\usage{
resample_long(x, statistic, boots, sigma = 150, nk_weight = FALSE,
  fws = NULL, n_snps = T, report = 10000, level = "group")
}
\arguments{
\item{x}{Input data containing \emph{unsmoothed} statistic of interest and SNP position. Note that this data must contain a column titled "group" containing the linkage group/chromosome/scaffold for each observed value, since positions are drawn from the same group only.}

\item{statistic}{A character string designating the \emph{unsmoothed} statistic to smooth.}

\item{boots}{Number of bootstraps to run.}

\item{sigma}{Size variable in kb for gaussian smoothing. Full window size is 6*sigma.}

\item{nk_weight}{If true, weights smoothing by randomly drawing variables from a column titled "nk" in x.}

\item{fws}{If using a window with a fixed slide rather than snp centralized, provide a data frame of window positions.}

\item{n_snps}{If true, draws all random numbers up front. A column containing the number of snps in each selected window must be provided, titled "n_snps", in either x or fws. This allows for faster processing. Otherwise, random numbers will be drawn after this value is determined for each bootstrap.}

\item{report}{Progress report interval, in bootstraps.}

\item{level}{Character or NULL, default "group." What level to bootstrap across?}
}
\value{
A numeric vector containing bootstrapped smoothed values.
}
\description{
\code{resample_long} creates a distribution of bootstrapped smoothed values for a desired statistic.
}
\details{
Bootstrapps are conducted as described by Hohenlohe et al. (2010), Population genomics of parallel adaptation in threespine stickleback using sequenced RAD tags.

For each bootstrap, this function draws random window position, then draws random statistics from all provided SNPs to fill each observed position on that window and calculates a smoothed statistic for that window using gaussian-smoothing.

This function can pick windows to bootstrap either from a provided data frame or around SNPs given a particular window size. It can also weight observations by random wieghts drawn from the provided SNP data. The function is built to run over one level, such as linkage group or chromosome.
}
\examples{
resample_long(randPI[randPI$pop == "A" & randPI$group == "chr1",], "pi", 100, 200, TRUE, randSMOOTHed[randSMOOTHed$pop == "A" & randSMOOTHed$group == "chr1",], TRUE, 10)

}
