% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping_functions.R
\name{resample_long_par}
\alias{resample_long_par}
\title{Runs \code{\link{resample_long}} in parallel across linkage groups/chromosomes.}
\usage{
resample_long_par(x, num_cores, statistic, boots, sigma = 150,
  nk_weight = FALSE, fws = NULL)
}
\arguments{
\item{x}{Input data containing \emph{unsmoothed} statistic of interest and SNP position.}

\item{num_cores}{The number of processing cores/nodes/threads.}

\item{statistic}{A character string designating the \emph{unsmoothed} statistic to smooth.}

\item{boots}{Number of bootstraps to run.}

\item{sigma}{Size variable in kb for gaussian smoothing. Full window size is 6*sigma.}

\item{nk_weight}{If true, weights smoothing by randomly drawing variables from a column titled "nk" in x.}

\item{fws}{If using a window with a fixed slide rather than snp centralized, provide a data frame of window positions.}
}
\value{
A numeric vector containing bootstrapped smoothed values.
}
\description{
\code{resample_long_par} runs \code{\link{resample_long}} in parallel across multiple linkage groups. To avoid read/write lag,this version does all random draws locally within each loop. Uses the doParallel and doSNOW packages.
}
\examples{
resample_long_par(randPI[randPI$pop == "A" & randPI$group == "chr1",], 3, "pi", 100, 200, TRUE, randSMOOTHed[randSMOOTHed$pop == "A" & randSMOOTHed$group == "chr1",])

}
