% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rgap_snps}
\alias{rgap_snps}
\title{Get random snps each window.}
\usage{
rgap_snps(x, gap, method = "gap", n = 1, levels = "group")
}
\arguments{
\item{x}{data.frame. Input data, must contain a column titled 'position' containing genomic positions of the SNPs.}

\item{gap}{Integer. How big are the windows from which to sample SNPs?}

\item{method}{Character, default "window". By which method should the SNPs be chosen?}

\item{n}{Integer, default 1. For "window" method. How many SNPs should be taken per window?}

\item{levels}{Character, default "group". Across which levels should the windows be partitioned (ex. "chromosome")?}
}
\value{
A data.frame containing subset SNPs.
}
\description{
\code{rgap_snps} grabs n random SNPs every window of a given length. Can do this across multiple linkage groups/chromosomes.
}
\details{
Possible methods:
\itemize{
   \item{window: }{Takes n random SNPs every window. This can result in SNPs that are much nearer to each other than the gap length.}
   \item{gap: }{Takes the first possible SNP, then takes the next SNP at least gap bases away. Non-random.}
}
}
