% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{row_specific_gsub}
\alias{row_specific_gsub}
\title{Perform a row-specific gsub/match given a key table}
\usage{
row_specific_gsub(x, lookup, lookup_match, patch = "_")
}
\arguments{
\item{x}{input data}

\item{lookup}{table of matches. Each column is something that can be gsub-ed
and each row corresponds to the rows of x}

\item{lookup_match}{a character vector with length equal to the number of columns
of lookup. In order, the matches in x for the contents of each column
in lookup.}

\item{patch}{A character string that will be used to join row ID to contents.
SHOULD NOT BE FOUND ANYWHERE IN X, LOOKUP, OR LOOKUP_MATCH.}
}
\description{
Perform a row-specific gsub/match given a key table
}
