% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony_interface.R
\name{run_colony}
\alias{run_colony}
\title{Create and infile, run, and gather some basic results using the COLONY parentage analysis program.}
\usage{
run_colony(x, colony_path = ".", outfile = "colony_input",
  method = "FPLS", run_length = 2, sibship_prior = 0,
  paternal_sib_size = NULL, maternal_sib_size = NULL, nruns = 1,
  seed = NULL, maternal_genotypes = NULL, paternal_genotypes = NULL,
  maternal_inclusion_prob = 0, paternal_inclusion_prob = 0,
  update_af = TRUE, dioecious = TRUE, inbreeding = TRUE,
  male_monogamous = FALSE, female_monogamous = FALSE,
  clone_inference = FALSE, sibship_scaling = TRUE, known_af = FALSE,
  precision = 2, dropout = 0.01, genotyping_error = 0.01,
  known_maternal_dryads = NULL, known_paternal_dryads = NULL,
  known_maternal_max_mismatches = 0, known_paternal_max_mismatches = 0,
  known_maternal_sibships = NULL, known_paternal_sibships = NULL,
  maternal_exclusions = NULL, paternal_exclusions = NULL,
  excluded_maternal_siblings = NULL, excluded_paternal_siblings = NULL)
}
\description{
Create a COLONY infile using snpRdata sets containg offspring and possibly paternal genotypes given
specified parameters. Requires that the COLONY program is installed locally. Input and output files
will be stored in a colony folder created in the current working directory.
}
