% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_enrichment.R
\name{run_enrichment}
\alias{run_enrichment}
\title{Determine SNP enrichment within genomic loci}
\usage{
run_enrichment(reg = NULL, ld.snps, matched.list)
}
\arguments{
\item{reg}{either a character vector to a BED file or a GRanges object}

\item{ld.snps}{list object created by \code{ld_SNPs()}}

\item{matched.list}{list of matched SNPs created by \code{matching_SNPs()}}
}
\description{
This function determines whether a set of SNPs are enriched, compared to 
matched SNPs, within a defined set of genomic regions.
}
