% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{run_genomic_prediction}
\alias{run_genomic_prediction}
\title{Interface with BGLR to run genomic prediction with snpRdata objects.}
\usage{
run_genomic_prediction(
  x,
  response,
  iterations,
  burn_in,
  thin,
  model = "BayesB"
)
}
\arguments{
\item{x}{snpRdata object}

\item{response}{character. Name of the column containing the response
variable of interest. Must match a column name in sample metadata.}

\item{iterations}{numeric. Number of iterations to run the MCMC chain for.}

\item{burn_in}{numeric. Number of burn in iterations to run prior to the MCMC
chain.}

\item{thin}{numeric. Number of iterations to discard between each recorded
data point.}

\item{model}{character, default "BayesB". Prediction model to use, see
description for the ETA argument in \code{\link[BGLR]{BGLR}}.}
}
\value{
A list containing: \itemize{ \item{model: } The model output from
  BGLR. See \code{\link[BGLR]{BGLR}}. \item{h2: } Estimated heritability of
  the response variable. \item{transposed_interpolated_genotypes: }
  Transposed genotypes, the input for BGLR. \item{predictions: } A data.frame
  containing the provided phenotypes and the predicted Breeding Values (BVs)
  for those phenotypes. }
}
\description{
Run genomic prediction given a single response variable (usually a phenotype)
using the \code{\link[BGLR]{BGLR}} function. Unlike other snpR functions,
this returns the resulting model direcly, so overwrite with caution.
}
\details{
This function is provided as a wrapper to plug snpRdata objects into the
\code{\link[BGLR]{BGLR}} function in order to easily run genomic prediction
on a simple model where a single, sample specific meta data variable is
provided as the response variable. To do so, this function formats the data
into a transposed "sn" format, as described in \code{\link{format_snps}}
using the bernoulli method to interpolate missing genotypes. Several
different prediction models are available, see the documentation the ETA
argument in \code{\link[BGLR]{BGLR}} for details. Defaults to the "BayesB"
model, which assumes a "spike-slab" prior for allele effects on phenotype
where most markers have a very small effect size and a few can have a much
larger effect.

Unlike most snpR functions, this function does not support facets, since each
run can be very slow. Instead, an individual facet and facet level of
interest should be selected with \code{\link{subset_snpR_data}}. See
examples.

See documentation for \code{\link[BGLR]{BGLR}} for more details and for a
full list of references.
}
\examples{
# run and plot a basic prediction
## add some dummy phenotypic data.
sample.meta <- cbind(weight = rnorm(ncol(stickSNPs)), stickSNPs@sample.meta)
dat <- import.snpR.data(as.data.frame(stickSNPs), stickSNPs@snp.meta, sample.meta)
## run prediction
gp <- run_genomic_prediction(dat, response = "weight", iterations = 1000, burn_in = 100, thin = 10)
## dummy phenotypes vs. predicted Breeding Values for those dummy predictions.
with(gp$predictions, plot(phenotype, predicted_BV))

}
\references{
Pérez, P., and de los Campos, G. (2014). \emph{Genetics}.
}
\author{
William Hemstrom
}
