% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association_functions.R
\name{run_random_forest}
\alias{run_random_forest}
\title{Run a RANGER random forest using snpRdata for a given phenotype or model.}
\usage{
run_random_forest(
  x,
  facets = NULL,
  response,
  formula = NULL,
  num.trees = 10000,
  mtry = NULL,
  importance = "impurity_corrected",
  interpolate = "bernoulli",
  pvals = TRUE,
  par = FALSE,
  ...
)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character, default NULL. Categorical metadata variables by
which to break up analysis. See \code{\link{Facets_in_snpR}} for more
details.}

\item{response}{character. Name of the column containing the response
  variable of interest. Must match a column name in sample metadata. Response
must be categorical, with only two categories.}

\item{formula}{charcter, default NULL. Model for the response variable, as
described in \code{\link[stats]{formula}}. If NULL, the model will be
equivalent to response ~ 1.}

\item{num.trees}{numeric, default 10000. Number of trees to grow. Higher
numbers will increase model accuracy, but increase calculation time. See
\code{\link[ranger]{ranger}} for details.}

\item{mtry}{numeric, default is the square root of the number of SNPs. Number
of variables (SNPs) by which to split each node. See
\code{\link[ranger]{ranger}} for details.}

\item{importance}{character, default "impurity_corrected". The method by
  which SNP importance is determined. Options: \itemize{\item{impurity}
\item{impurity_corrected} \item{permutation}}. See
\code{\link[ranger]{ranger}} for details.}

\item{interpolate}{character, default "bernoulli". Interpolation method for
missing data. Options: \itemize{\item{bernoulli: }binomial draws for the
minor allele. \item{af: } insertion of the average allele frequency}.}

\item{pvals}{logical, default TRUE. Determines if pvalues should be
calculated for importance values. If the response variable is quantitative,
no pvalues will be returned, since they must be calculated via permutation
and this is very slow. For details, see
\code{\link[ranger]{importance_pvalues}}.}

\item{par}{numeric, default FALSE. Number of parallel computing cores to use
for computing RFs across multiple facet levels or within a single facet if
only a single category is run (either a one-category facet or no facet).}

\item{...}{Additional arguments passed to \code{\link[ranger]{ranger}}.}
}
\value{
A list containing: \itemize{\item{data: } A snpRdata object with RF
  importance values merged in to the stats slot. \item{models: } A named list
containing both the models and data.frames containing the predictions vs
observed phenotypes.}
}
\description{
Creates forest machine learning models using snpRdata objects via interface
with \code{\link[ranger]{ranger}}. Models can be created either for a
specific phenotype with no-covariates or using a formula which follows the
basic format specified in \code{\link{formula}}.
}
\details{
Random forest models can be created across multiple facets of the data at
once following the typical snpR framework explained in
\code{\link[snpR]{Facets_in_snpR}}. Since RF models are calculated without
allowing for any SNP-specific categories (e.g. independent of chromosome
etc.), any sample level facets provided will be ignored. As usual, if facets
is set to NULL, an RF will be calculated for all samples without splitting
across any sample metadata categories.

Since the \code{\link[ranger]{ranger}} RF implementation can behave
unexpectedly when given incomplete data, missing genotypes will be imputed.
Imputation can occur either via the insertion of the average allele frequency
or via binomial draws for the minor allele using the "af" or "bernoulli"
options for the "interpolate" argument.

Extra arguments can be passed to \code{\link[ranger]{ranger}}.

In general, random forest parameters should be tuned so as to reduce the out-of-bag error rates (OOB-ER).
This value is visible in the returned object under the model lists. Simply calling a specific model will
output the OOB-ER, and they are also stored under the 'prediction.error' name in the model. For details
on tuning RF models, we recommend Goldstein et al. (2011).

For more detail on the random forest model and ranger arguments, see
\code{\link[ranger]{ranger}}.
}
\examples{
#' # run and plot a basic rf
## add some dummy phenotypic data.
dat <- stickSNPs
sample.meta(dat) <- cbind(weight = rnorm(ncol(stickSNPs)), sample.meta(stickSNPs))
## run rf
rf <- run_random_forest(dat, response = "weight")
rf$models
## dummy phenotypes vs. predicted 
with(rf$models$.base_.base$predictions, plot(pheno, predicted)) # not overfit


}
\references{
Wright, Marvin N and Ziegler, Andreas. (2017). ranger: A Fast
  Implementation of Random Forests for High Dimensional Data in C++ and R.
  \emph{Journal of Statistical Software}.

Goldstein et al. (2011). Random forests for genetic association
  studies. \emph{Statistical Applications in Genetics and Molecular Biology}.
}
\seealso{
\code{\link[ranger]{ranger}} \code{\link[ranger]{predict.ranger}}.
}
\author{
William Hemstrom
}
