% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequoia_interface.R
\name{run_sequoia}
\alias{run_sequoia}
\title{Run Sequoia with snpR}
\usage{
run_sequoia(
  x,
  facets = NULL,
  run_dupcheck = T,
  run_parents = T,
  run_pedigree = T,
  pMaxSibIter = 10,
  min_maf = 0.3,
  min_ind = 0.5,
  ...
)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{FALSE or character, default FALSE. Sample-specific facets over which the sequoia is called to run.}

\item{run_dupcheck}{FALSE or TRUE, default FALSE. Should a duplicate check be run on this dataset?}

\item{run_parents}{FALSE or TRUE, default FALSE. Should parentage assignments be run with the dataset?}

\item{run_pedigree}{FALSE or TRUE, default FALSE. Should a pedigree be constructed with the dataset? Requires run_parents to have been completed first.}
}
\value{
A dataframe for each facet specified with sequoia output summary information.
}
\description{
\code{run_sequoia} a tool for parentage assignment and pedigree construction. snpR integrates a program and package written by Jisca Huisman \itemize{\item{paper: } https://doi.org/10.1111/1755-0998.12665 \item{github page: } https://github.com/JiscaH/sequoia}. Note that this function \emph{is not overwrite safe!}
}
\details{
Possible options: \itemize{ \item{facets: }{Specification for grouping of data for which analyses should be ran independently} \item{run_dupcheck: }{Sequoia function used to check for duplicate samples in the dataset. It is reccomended to drop duplicated samples from the analysis. } \item{run_parents: }{Sequoia function to assign parents. This runs quickly and is required before running the run_pedigree command. } \item{run_pedigree: }{Sequoia function to construct full pedigree for the sample set. Sibship clustering takes a lot of time. }}

Note that there are many more Sequoia specific arguments that can be added to change from the default settings (eg. ErrorM, Tassign, Tfilt, GetMaybeRel, etc.) See documentation for \code{\link[sequoia]{sequioa}}.
}
\examples{
# to follow an example using the stickSNPs example dataset you need to add some variables that don't exist in the actual dataset. 
a <- 2013:2015 #create a vector of possible birthyears
b <- c("M", "F", "U") #create a vector of possible sexes
stk <- stickSNPs
sample.meta(stk)$BirthYear <- sample(x = a, size = nrow(stk@sample.meta), replace = T) #create birthyears
sample.meta(stk)$ID <- 1:nsamps(stk) #create unique sampleID
sample.meta(stk)$Sex <- sample(x= b, size = nsamps(stk), replace = T) # create sexes
dup <- run_sequoia(x = stk, run_dupcheck = T, run_parents = F, run_pedigree = F)
}
\author{
William Hemstrom

Melissa Jones
}
