% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{sanity_check_window}
\alias{sanity_check_window}
\title{Do sanity checks for many window specific functions in snpR.}
\usage{
sanity_check_window(
  x,
  sigma,
  step,
  stats.type,
  nk,
  facets,
  stats = NULL,
  good.types = NULL
)
}
\arguments{
\item{x}{snpRdata object.}

\item{sigma}{numeric. Window size, in kilobases.}

\item{step}{numeric or NULL. Step size, in kilobases}

\item{stats.type}{character, either "single" or "pairwise". The type of stats
being checked, see documentation for \code{\link{calc_smoothed_averages}}.}

\item{nk}{Logical. Determines if nk values are to be used.}

\item{facets}{character or NULL. Defines facets to run, following typical
rules as described in \code{\link{Facets_in_snpR}}.}

\item{stats}{character or NULL, default NULL. Statistics (pi, etc.) used,
really only for bootstrapping.}

\item{good.types}{character or NULL, default NULL. Good statistics, for use
with the stats arguement.}
}
\description{
Internal function only. Since many window related functions have similar
sanity checks, they have been integrated here. For examples, see use in
functions like \code{\link{do_bootstraps}}.
}
\author{
William Hemstrom
}
