% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata_methods.R
\docType{methods}
\name{show,snpRdata-method}
\alias{show,snpRdata-method}
\title{Display snpRdata objects}
\usage{
\S4method{show}{snpRdata}(object)
}
\arguments{
\item{object}{A \code{snpRdata} object.}
}
\description{
Display snpRdata objects
}
