% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{SingleStats}
\alias{SingleStats}
\alias{calc_pi}
\alias{calc_hwe}
\alias{calc_ho}
\alias{calc_private}
\alias{calc_maf}
\title{Calculate standard single-SNP genetic statistics}
\usage{
calc_pi(x, facets = NULL)

calc_maf(x, facets = NULL)

calc_ho(x, facets = NULL)

calc_private(x, facets = NULL)

calc_hwe(x, facets = NULL, method = "exact")
}
\arguments{
\item{x}{snpRdata. Input SNP data.}

\item{facets}{character. Facets to use.}

\item{method}{character, default "exact". Either exact or chisq. Defines the
method to use for calculating p-values for HWE divergence.}
}
\value{
snpRdata object with requested stats merged into the stats socket
}
\description{
These functions calculate a basic genetic statistics from SNP data contained
in snpRdata objects, splitting samples by any number of provided facets. Since
these statistics all use only a single SNP at a time, they ignore any SNP
specific facet levels.
}
\details{
The data can be broken up categorically by sample metadata, as
described in \code{\link{Facets_in_snpR}}.
}
\section{Functions}{
\itemize{
\item \code{calc_pi}: pi (average number of pairwise differences/expected heterozygosity)

\item \code{calc_maf}: minor allele frequency

\item \code{calc_ho}: observed heterozygosity

\item \code{calc_private}: find private alleles

\item \code{calc_hwe}: p-values for Hardy-Wienberg Equilibrium divergence
}}

\section{pi}{


 Calculates pi (genetic diversity/average number of pairwise differences)
 according to Hohenlohe et al. (2010).
}

\section{maf}{


 Calculates minor allele frequencies and note identities and counts of major
 and minor alleles.
}

\section{ho}{


 Calculates observed heterozygosity.
}

\section{private alleles}{


 Determines if each SNP is a private allele across all levels in each sample
 facet. Will return an error of no sample  facets are provided.
}

\section{HWE}{


 Calculates a p-value for the null hypothesis that a population is in HWE at
 a given locus. Several methods available:
 \itemize{
   \item{"exact"} Exact test according to Wigginton, JE, Cutler, DJ,
                and Abecasis, GR (2005). Slightly slower.
   \item{"chisq"} Chi-squared test. May produce poor results when sample
                sizes for any observed or expected genotypes are small.
 }

 For the exact test, code derived from \url{http://csg.sph.umich.edu/abecasis/Exact/snp_hwe.r}
}

\examples{
# base facet
x <- calc_pi(stickSNPs)

# multiple facets
x <- calc_pi(stickSNPs, facets = c("pop", "pop.fam"))

# HWE
x <- calc_HWE(stickSNPs, facets = c("pop", "pop.fam"), method = "exact")
}
\references{
Wigginton, JE, Cutler, DJ, and Abecasis, GR (2005). \emph{American Journal of Human Genetics}

Hohenlohe et al. (2010). \emph{PLOS Genetics}
}
\author{
William Hemstrom
}
