% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{smart.merge}
\alias{smart.merge}
\title{Merging function used in merge.snpR.stats and elsewhere}
\usage{
smart.merge(n.s, o.s, meta.names, starter.meta)
}
\arguments{
\item{n.s}{new stats}

\item{o.s}{old stats}

\item{meta.names}{names of the metadata columns, usually everything up to .snp.id}

\item{starter.meta}{any metadata columns that should specifically be put at the start of the output data (such as facet, subfacet, facet.type)}
}
\description{
Merging function used in merge.snpR.stats and elsewhere
}
