% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{snpR.stats}
\alias{snpR.stats}
\title{SNP data analysis with snpR}
\usage{
snpR.stats(x, ecs, stats = "basic", filter = FALSE, smooth = FALSE,
  bootstrap = FALSE, graphs = FALSE, sigma = NULL, ws = NULL,
  pop = NULL, nk = TRUE, input = "NN", mDat = "N", maf = 0.05,
  hf_hets = 0.55, bi_al = TRUE, non_poly = TRUE, min_ind = FALSE,
  min_loci = FALSE, pop_maf = FALSE, filt_re = "partial",
  Fst_method = "genepop", LD_level = c("group", "pop"), LD_chr = "all",
  LD_g_par = FALSE, tsd_level = c("group", "pop"),
  smooth_levs = c("group", "pop"), boot_levs = c("pop"),
  p.alt = "two-sided", bootstrap_par = FALSE, num_boots = NULL,
  par_cores = FALSE)
}
\arguments{
\item{x}{Data frame. Input data, in either numeric or character formats, as given by format_snps output options 2 or 6.}

\item{ecs}{Numeric value. Number of extra metadata columns at the start of x.}

\item{stats}{Character vector, default "basic" (see description). Which stats should be calculated?}

\item{filter}{Logical, default FALSE. Should the SNP data be filtered?}

\item{smooth}{Character vector or FALSE, default FALSE. Should variables be smoothed? If so, which?}

\item{bootstrap}{Character vector or FALSE, default FALSE. Should bootstraps of smoothed windows be conducted? If so, which variables?}

\item{graphs}{Character vector or FALSE, default FALSE. Which variables should be plotted? Plots \emph{smoothed} values of most statistics save population structure, LD, and Tajima's D. As such, requires the statistics to be set via the "smooth" argument.}

\item{sigma}{Numeric, default NULL. Smoothing statistic/window size, in kb.}

\item{ws}{Numeric, default NULL.  How far should the window slide when smoothing? If null, a window is produced centered at every SNP.}

\item{pop}{Table of length > 1, default NULL. Table containing the number of samples in each population. Samples should be sorted by population and ordered identically in data.}

\item{nk}{Logical, default TRUE. Should contributions to smoothed windows be weighted by the number of called alleles at each site?}

\item{input}{Character string, default "NN". How are genotypes noted? Supports "NN" and "0000". See \code{\link{format_snps}}.}

\item{mDat}{Character string, default "N". How are missing \emph{alleles} coded in x (typically "N" or "00")?}

\item{maf}{FALSE or numeric between 0 and 1, default FALSE. Minimum acceptable minor allele frequency during filtering.}

\item{hf_hets}{FALSE or numeric between 0 and 1, default 0.55 Maximum acceptable heterozygote frequency during filtering.}

\item{bi_al}{Logical, default TRUE. Should non-biallelic SNPs be removed during filtering?}

\item{non_poly}{Logical, default TRUE. Should non-polymorphic loci be removed during filtering?}

\item{min_ind}{FALSE or integer, default FALSE. Minimum number of individuals in which a loci must be sequenced during filtering.}

\item{min_loci}{FALSE or numeric between 0 and 1, default FALSE. Minimum proportion of SNPs an individual must be genotyped at to be kept during filtering.}

\item{pop_maf}{Logical, default FALSE. For filtering, should loci be kept if they are above the minor allele frequency cuttoff in \emph{any} population?}

\item{filt_re}{FALSE, "partial", or "full", default "partial". How should loci be re_filtered after individuals are filtered? See \code{\link{filter_snps}}.}

\item{Fst_method}{Character vector, default "genepop". Which FST estimator should be used? See documentation for options.}

\item{LD_level}{Character vector or FALSE, default c("group", "pop"). Names of metadata columns by which to split the data for LD calculations.}

\item{LD_chr}{Character vector, default "all". Names of linkage groups/chromsomes for which pairwise LD values should be calculated at all SNPs. Identifiers must be contained in a column named either "group" or "chr".}

\item{LD_g_par}{Logical, default NULL. Shoud LD be done in parallel across multiple linkage groups/chromsomes?}

\item{tsd_level}{Character vector or FALSE, default c("group", "pop"). Names of metadata columns by which to split the data for Tajima's D calculations.}

\item{smooth_levs}{Character vector or FALSE, default c("group", "pop"). Names of metadata columns by which to split the data for smoothing.}

\item{boot_levs}{Character vector or FALSE, default c("group", "pop"). Names of metadata columns by which to split the data for bootstraps.}

\item{p.alt}{Character string, default "two-sided". If bootstraps are done, which alternative hypothesis should be used? See documentation for options.}

\item{num_boots}{Numeric, default NULL. If bootstraps are done, how many should be performed per stat?}

\item{par_cores}{Numeric, default FALSE. If parallel processing is requested, how many cores should be used?}

\item{bootstap_par}{Logical, default FALSE. If bootstraps are done, should they be done in parallel?}
}
\value{
A named list of objects corresponding to requested statistics/graphs/ect.
}
\description{
\code{snpR.stats} calls many snpR functions based on the provided arguments to do the majority of the basic statistical process required for SNP analysis.
}
\details{
Description of x:
   Contains metadata in columns 1:ecs. Remainder of columns contain genotype calls for each individual. Each row is a different SNP, as given by format_snps output options 4 or 6. Requires the column containing the position of the loci in base pairs be named "position". If population info is given, it must be contained in a column named "pop".
}
\examples{

}
