% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata.R
\name{snpRdata_dims}
\alias{snpRdata_dims}
\alias{nsnps}
\alias{nrow,snpRdata-method}
\alias{nsamps}
\alias{ncol,snpRdata-method}
\alias{dim,snpRdata-method}
\title{Get the number of samples in a snpRdata object.}
\usage{
nsnps(x)

\S4method{nrow}{snpRdata}(x)

nsamps(x)

\S4method{ncol}{snpRdata}(x)

\S4method{dim}{snpRdata}(x)
}
\arguments{
\item{x}{snpRdata object}
}
\description{
Get the number of samples in a snpRdata object.
}
\section{Functions}{
\itemize{
\item \code{nsnps}: get the number of SNPs

\item \code{nrow,snpRdata-method}: identical to nsnps

\item \code{nsamps}: get the number of samples

\item \code{ncol,snpRdata-method}: identical to nsamps

\item \code{dim,snpRdata-method}: get the number of SNPs and samples
}}

