% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{stickRAW}
\alias{stickRAW}
\title{Deschutes stickleback raw data.}
\format{
An object of class \code{data.frame} with 833 rows and 423 columns.
}
\usage{
stickRAW
}
\description{
A dataset containing genotypes at 833 SNP loci for 420 Three-spined
Stickleback collected in 2015 from several locations in the Deschutes River,
Oregon, USA. For more information, see Hemstrom et al (in prep).
}
\details{
Genotypes noted as two characters representing bases sequenced at a specific
site (e.g. "CC" or "GT"). Each SNP occupies one row, each sample occupies on
column. Missing data coded as "NN".

Column names refer to the population of origin (first three characters) and
the sample ID.

The first three columns contain snp specific meta data: \itemize{ \item{snp: }
SNP ID. \item{chr: } Linkage group, essentially chromosome. \item{position:
} Position of the SNP along the linkage group, in base pairs. }
}
\references{
Hemstrom et al (in prep)
}
\keyword{datasets}
