% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{subset_snpR_data}
\alias{subset_snpR_data}
\title{Subset snpRdata objects}
\usage{
subset_snpR_data(x, snps = 1:nrow(x), samps = 1:ncol(x),
  facets = NULL, subfacets = NULL, snp.facets = NULL,
  snp.subfacets = NULL)
}
\arguments{
\item{x}{snpRdata object.}

\item{snps}{numeric, default \code{1:nrow(x)}. Row numbers corresponding to
SNPs to keep.}

\item{samps}{numeric, default \code{1:ncol(x)}. Column numbers corresponding
to samples to keep.}

\item{facets}{character, default NULL. \emph{sample specific} facets over
which subsetting will occur.}

\item{subfacets}{character, default NULL. Levels of the specified sample level
facet to keep. Samples in other levels will be removed.}

\item{snp.facets}{characet, default NULL. \emph{snp specific} facets over
which subsetting will occur.}

\item{snp.facets}{character, default NULL. Levels of the specified snp facets
to keep. SNPs in other levels will be removed.}
}
\description{
Subsets snpRdata objects by specific snps, samples, facets, subfacets, ect.
Throws away as few calculated stats as possible.
}
\details{
This function exists to intelligently subset snpRdata. While the typical
bracket notation can be used to subset snpRdata, that method is inherited from
the data.frame S3 object class for consistancy. As such, the result will be a
data.frame, not a snpRdata object. This function keeps the data in a snpRdata
object, and retains as much previously calculated data as possible.

If \emph{samples} are removed, most statistics will be thrown away, since
values like pi will change. In contrast, if \emph{snps} are removed, many
statistics will simply be subset. LD, bootstraps, and window stats will
\emph{always} be discarded, so overwrite with caution.

Sample and snp facets to subset over can be provided. Facet levels to keep are
provided in the corresponding subfacet arguments. Facets designated as
described in \code{\link{Facets_in_snpR}}.
}
\examples{
# Keep only individuals in the ASP and PAL populations and on the LGIX or LGIV chromosome.
subset_snpR_data(stickSNPs, facets = "pop", subfacets = c("ASP", "PAL"), snp.facets = "group", snp.subfacets = c("groupIX", "groupIV"))

# Keep only individuals and SNPs 1 through 10
subset_snpR_data(stickSNPs, snps = 1:10, samps = 1:10)

# Keep SNPs 1:100, individuals in the ASP population
subset_snpR_data(stickSNPs, snps = 1:100, facets = "pop", subfacets = "ASP")
}
