% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata_methods.R, R/utility_functions.R
\docType{methods}
\name{[,snpRdata,ANY,ANY,ANY-method}
\alias{[,snpRdata,ANY,ANY,ANY-method}
\alias{[,snpRdata-method}
\alias{subset_snpRdata}
\alias{subset_snpR_data}
\alias{[}
\title{Subset snpRdata objects}
\usage{
\S4method{[}{snpRdata,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

subset_snpR_data(x, .snps = 1:nsnps(x), .samps = 1:nsamps(x), ...)
}
\arguments{
\item{x}{snpRdata object.}

\item{i}{numeric, default \code{1:nrow(x)}. Row numbers corresponding to
SNPs to keep. Negative subscripts allowed.}

\item{j}{numeric, default \code{1:nrow(x)}. Row numbers corresponding to
SNPs to keep. Negative subscripts allowed.}

\item{...}{Facet subsetting may be specified by providing the facet as an 
argument and then providing the levels to keep or remove. Setting
pop = "ASP", for example, would keep only samples in the "ASP" level of the
pop facet, and setting pop.fam = "ASP.A" would keep only samples the ASP
pop and the A fam. Negative subscripts are allowed: pop = -c("ASP", "PAL")
would remove samples in the ASP or PAL pops. Subsetting by 
multiple facets is supported, although negative and positive subscripts
cannot be mixed across sample or SNP facets. They may be mixed between the
two.}

\item{drop}{logical, default FALSE. Deprecated.}

\item{.snps}{numeric, default \code{1:nrow(x)}. Row numbers corresponding to
SNPs to keep. Negative subscripts allowed.}

\item{.samps}{numeric, default \code{1:ncol(x)}. Column numbers corresponding
to samples to keep. Negative subscripts allowed.}
}
\description{
Subsets snpRdata objects by specific snps, samples, facets, subfacets, etc.
Statistics will need to be recalculated after subsetting to avoid confusion. 
The bracket operators can be alternatively, following the same syntax.
}
\details{
Sample and snp facets to subset over can be provided. Filtering by facet
categories is performed by naming the facet as an argument name, then
providing the levels to keep or remove to that argument. See examples. Facets
designated as described in \code{\link{Facets_in_snpR}}.
}
\section{Functions}{
\itemize{
\item \code{x[i}: extraction operator

\item \code{subset_snpR_data()}: subset_snpR_data

}}
\examples{
# Keep only individuals in the ASP and PAL populations 
# and on the LGIX or LGIV chromosome.
subset_snpR_data(stickSNPs, pop = c("ASP", "PAL"), 
                 chr = c("groupIX", "groupIV"))
                 
# keep individuals/SNPs in the first 10 columns/rows.
subset_snpR_data(stickSNPs, 1:10, 1:10)

# negative subscripts: remove individuals in family B
subset_snpR_data(stickSNPs, fam = -"B")

# negative subscripts: remove individuals in pop PAL AND fam B or C, 
# and keep only SNPs on LGIV
subset_snpR_data(stickSNPs, pop.fam = -c("PAL.B", "PAL.C"),
                 chr = "groupIV")
                 
# using the bracket operator, same as example 1
stickSNPs[pop = c("ASP", "PAL"), chr = c("groupIX", "groupIV")]

# bracket operator, excluding first ten SNPs and only keeping pop = PAL,
# fam = B
stickSNPs[-c(1:10), pop.fam = c("PAL.B")]
}
