% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{summarize_facets}
\alias{summarize_facets}
\title{Summarize possible snpRdata object facet options}
\usage{
summarize_facets(x, facets = NULL)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{character. Categorical metadata variables by which to break up
analysis. See \code{\link{Facets_in_snpR}} for more details. If NULL, the
possible SNP and sample facets will be listed. If facets are instead
provided, the \emph{categories} for each facet will instead be listed.}
}
\value{
A named list containing either the possible SNP and sample facets or
  the categories for all of the requested facets.
}
\description{
List either all of the possible SNP and sample facets (if called with no
facets) or all of the categories for each requested facet.
}
\details{
If called with sample facets, returns a table of counts of each category
otherwise returns a vector of categories.
}
\examples{
# list available facets
summarize_facets(stickSNPs)

# return details for a few facets
summarize_facets(stickSNPs, c("pop", "chr.pop", "fam.pop"))
}
\author{
William Hemstrom
}
