% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_growth.R
\name{survival_distributions}
\alias{survival_distributions}
\alias{normal_survival}
\alias{BL_survival}
\title{Calculate survivability probabilities.}
\usage{
normal_survival(phenotypes, opt_pheno, var, max.survival = 0.5, ...)

BL_survival(phenotypes, opt_pheno, omega, ...)
}
\arguments{
\item{phenotypes}{numeric. Vector of phenotypes.}

\item{opt_pheno}{numeric. Optimum phenotype.}

\item{var}{numeric. Variance of the surivial distribution.}

\item{max.survival}{numeric. Maximum possible survival probability.}

\item{omega}{numeric. Strength of stabilizing selection. Smaller numbers mean
stronger selection.}
}
\value{
A vector of survival probabilities for each phenotype.
}
\description{
For a given optimum phenotype and surivival variance, calculate survival
probabilities given a specific distribution. For use in
\code{\link{simulate_populations}}.
}
\details{
Sevaral survival probablity distribution curves are provided. Each are coded
such that they expect a vector of phenotypes as the first argument and the
optimum phenotype as the second, with other arguments following, ending in
the `...` argument so that additional, unused arguments can be passed without
error. Other distributions can be coded in the same way if needed; call
\code{normal_survival} without parentheses to see an example.

Options:
\itemize{\item{normal_survival: }{Probabilities are given by
\code{\link[stats]{pnorm}}. For values above the
optimum phenotype, the surivial is given by 1 - the resulting value. If a
maxium survival other than 0.5 is given, the surivivals are scaled such that
individuals with the optimum phenotype will have the provided survival
probability.}
\item{BL_survival: } {Survival according to Burger and Lynch's (1995) 
equation 1.}}
}
\section{Functions}{
\itemize{
\item \code{normal_survival()}: normally distributed survival probabilities

\item \code{BL_survival()}: Burger and Lynch (1995) distributed
survival probabilities

}}
