% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{tabulate_genotypes}
\alias{tabulate_genotypes}
\title{Tabulate allele and genotype counts at each locus.}
\usage{
tabulate_genotypes(xv, snp_form, mDat, nsamp, nloci, verbose = F)
}
\arguments{
\item{xv}{Character vector containing genotype calls, ordered by locus and then individual, as produced by as.vector(t(data))}

\item{snp_form}{Integer. How many characters are there per genotype?}

\item{mDat}{Character string. How are missing \emph{genotypes} noted?}

\item{nsamp}{Integer. How many individuals does the data have?}

\item{nloci}{Integer. How many loci does the data have?}

\item{verbose}{Logical. Should the function report progress?}
}
\value{
A list of matrices. gs is the genotype matrix, as is the allele matrix, and wm is the genotype matrix with missing genotypes.
}
\description{
\code{tabulate_genotypes} creates matricies containing counts of observed alleles and genotypes at each locus.
}
\details{
This function is pirmarily used interally in several other funcitons, but may occasionally be useful.
}
\examples{
xv <- as.vector(t(stickSNPs[,c(4:ncol(stickSNPs))]))
tabulate_genotypes(xv, 2, "NN", ncol(stickSNPs) - 3, nrow(stickSNPs))

}
