% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony_interface.R
\name{write_colony_input}
\alias{write_colony_input}
\title{Write an input for the COLONY pedigree assignment program.}
\usage{
write_colony_input(
  x,
  outfile = "colony_input",
  method = "FPLS",
  run_length = 2,
  sampleIDs = NULL,
  sibship_prior = 0,
  paternal_sib_size = NULL,
  maternal_sib_size = NULL,
  nruns = 1,
  seed = NULL,
  maternal_genotypes = NULL,
  paternal_genotypes = NULL,
  maternal_inclusion_prob = 0,
  paternal_inclusion_prob = 0,
  update_af = TRUE,
  dioecious = TRUE,
  inbreeding = TRUE,
  male_monogamous = FALSE,
  female_monogamous = FALSE,
  clone_inference = FALSE,
  sibship_scaling = TRUE,
  known_af = FALSE,
  precision = 2,
  dropout = 0.01,
  genotyping_error = 0.01,
  known_maternal_dyads = NULL,
  known_paternal_dyads = NULL,
  known_maternal_max_mismatches = 0,
  known_paternal_max_mismatches = 0,
  known_maternal_sibships = NULL,
  known_paternal_sibships = NULL,
  maternal_exclusions = NULL,
  paternal_exclusions = NULL,
  excluded_maternal_siblings = NULL,
  excluded_paternal_siblings = NULL
)
}
\arguments{
\item{x}{snpRdata object containing offspring genotypes.}

\item{outfile}{character, default "colony_input". Output file name. A file
path may be provided (e.g. "colony/colony_run_1.txt").}

\item{method}{character, default "FPLS". Pedigree reconstruction method.
Options: \itemize{\item{"FLPS": }{Pure pairwise likelihood method, combines the full likelihood and pairwise likelihood methods. A good compromise between speed and accuracy.}
\item{"PLS": }{} }}

\item{run_length}{numeric in 1:4, default 2. Length of run, short/medium/long/verylong.}

\item{sibship_prior}{numeric in c(0, 0.25, 0.5, 1), default 0. Strength the sibship size prior (no prior, weak, medium, or strong). Values other than 0 require additional parameters.}

\item{paternal_sib_size}{numeric, default NULL. Minimum value is 0. The number of offspring in the candidate pool that are known to share the same father. If this value is not zero, then you must include a file with the known paternal sibship/paternity.}

\item{nruns}{integer, default 1. A number of replicate runs for the dataset.}

\item{seed}{integer, default NULL. Supply a four digit integer (eg: 1234, 9876) as a starting point for the algorithm.}

\item{maternal_genotypes}{snpRdata object containing maternal genotypes.}

\item{paternal_genotypes}{snpRdata object containing paternal genotypes.}

\item{maternal_inclusion_prob}{numeric, default 0. Probability ranging from 0 to 1.}

\item{paternal_inclusion_prob}{numeric, default 0. Probability ranging from 0 to 1.}

\item{update_af}{character, default TRUE. Should Colony update the allele frequencies used in the calculations?}

\item{dioecious}{character, default TRUE. Is this species diploid/dioecious?}

\item{inbreeding}{character, default TRUE. Should Colony assume inbreeding in the calculations?}

\item{male_monogamous}{character, default FALSE. Should Colony assume males are monogamous?}

\item{clone_inference}{character, default FALSE. Should Colony infer clones in the sample set?}

\item{sibship_scaling}{character, default TRUE. Should Colony scale sibling groups?}

\item{known_af}{character, default FALSE. If TRUE supply a vector of known allele frequencies.}

\item{precision}{integer in c(1,2,3,), default 2. Low/Medium/High/Very High for calculating the max likelihood.}

\item{dropout}{numeric in 0:1, default 0.01. Supply a flatrate value for all markers, or a vector corresponding to the allelic droput rate for each marker.}

\item{genotyping_error}{numeric in 0:1, default 0.01. Supply a flatrate value for all markers, or a vector corresponding to the genotyping error rate for each marker.}

\item{known_maternal_dyads}{character, default NULL. Supply list of known maternal-offspring dyads. Offspring ID in column 1, Maternal ID in column 2.}

\item{known_paternal_dyads}{character, default NULL. Supply list of known paternal-offspring dyads. Offspring ID in column 1, Paternal ID in column 2.}

\item{sampleIDS}{character, default NULL. Name of a column in the sample metadata that designates sample identifications/"names". Each name must be unique!}

\item{maternal_sibship_size}{numeric, default NULL. Minimum value is 0. The number of offspring in the candidate pool that are known to share the same mother. If this value is not zero, then you must include a file with the known paternal sibship/maternity.}

\item{female_monagamous}{character, default FALSE. Should Colony assume females are monogamous?}
}
\description{
Creates an input file in the format required for command-line usage of the
COLONY pedigree program. Requires a snpRdata object containing offspring
genotypes and can optionally take a snpRdata object containing maternal or
paternal genotypes, or both.
}
\author{
William Hemstrom

Melissa Jones
}
