% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_genetic_distances}
\alias{calc_genetic_distances}
\title{Calculate genetic distances between individuals or groups of samples.}
\usage{
calc_genetic_distances(
  x,
  facets = NULL,
  method = "Edwards",
  interpolate = "bernoulli"
)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character or NULL, default NULL. Facets for which to calculate
genetic distances, as described in \code{\link{Facets_in_snpR}}. If snp or
base level facets are requested, distances will be between individuals.
Otherwise, distances will be between the levels of the sample facets.}

\item{method}{character, default "Edwards". Name of the method to use.
Options: \itemize{\item{Edwards} Angular distance as described in Edwards
1971.} See details.}

\item{interpolate}{character, default "bernoulli". Missing data interpolation
method, solely for individual/individual distances. Options detailed in
documentation for \code{\link{format_snps}}.}
}
\value{
An overwrite-safe snpRdata object with genetic distance information
  (a named, nested list containing distance measures named according to
  facets and facet levels) added.
}
\description{
Calculates the genetic distances between either individuals or the levels of
any sample specific facets, broken apart by any requested snp level facets.
For details on methods, see details.
}
\details{
If a sample facet is requested, distances are calulcated via code adapted
from code derived from \code{\link[adegenet]{adegenet}}. Please cite them
alongside the tree-building and distance methods Available methods:
\itemize{\item{Edwards: } Angular distance as described in Edwards 1971.
\item{Nei: } Nei's (1978) genetic distance measure.}

Otherwise, genetic distances are calculated via the \code{\link[stats]{dist}}
function using genotypes formatted numerically (the "sn" option in
\code{\link{format_snps}}). In all cases, missing genotypes are first
interpolated according to the method provided to the 'interpolate' argument.
}
\examples{
# by pop:
y <- calc_genetic_distances(stickSNPs, facets = "pop", method = "Edwards")
get.snpR.stats(y, "pop", "genetic_distance")

# by chr and pop jointly
y <- calc_genetic_distances(stickSNPs, facets = "pop.chr", 
                            method = "Edwards")
get.snpR.stats(y, "pop.chr", "genetic_distance")

\dontrun{
# by pop and fam seperately
y <- calc_genetic_distances(stickSNPs, facets = c("pop", "fam"), 
                            method = "Edwards")
get.snpR.stats(y, c("pop", "chr"), "genetic_distance")

# individuals across all snps + plot
y <- calc_genetic_distances(stickSNPs)
dat <- get.snpR.stats(y, stats = "genetic_distance")$.base$.base$Edwards
heatmap(as.matrix(dat))
}
}
\references{
Edwards, A. W. F. (1971). Distances between populations on the
  basis of gene frequencies. Biometrics, 873-881.

Nei, M. (1978). Estimation of average heterozygosity and genetic
  distance from a small number of individuals. Genetics, 23, 341-369.

Jombart, T. (2008). adegenet: a R package for the multivariate
  analysis of genetic markers Bioinformatics 24: 1403-1405.
}
\author{
William Hemstrom
}
