% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_genetic_distances}
\alias{calc_genetic_distances}
\title{Caculate genetic distances between individuals or groups of samples.}
\usage{
calc_genetic_distances(
  x,
  facets = NULL,
  method = "Edwards",
  interpolate = "bernoulli"
)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character or NULL, default NULL. Facets for which to calculate
genetic distances, as described in \code{\link{Facets_in_snpR}}. If snp or
base level facets are requested, distances will be between individuals.
Otherwise, distances will be between the levels of the sample facets.}

\item{method}{character, default "Edwards". Name of the method to use.
Options: \itemize{\item{Edwards} Angular distance as described in Edwards
1971.} See details.}

\item{interpolate}{character, default "bernoulli". Missing data interpolation
method, solely for individual/individual distances. Options detailed in
documentation for \code{\link{format_snps}}.}
}
\value{
An overwrite-safe snpRdata object with genetic distance information
  (a named, nested list containing distance measures named according to
  facets and facet levels) added.
}
\description{
Calculates the genetic distances between either individuals or the levels of
any sample specific facets, broken apart by any requested snp level facets.
For details on methods, see details.
}
\details{
Available methods: \itemize{\item{Edwards} Angular distance as described in
Edwards 1971.}

All methods first calculate allele frequency matrices using
\code{\link{tabulate_allele_frequency_matrix}}, then use these matrices (hereafter
x) to calculate genetic distance.

Method details: \itemize{\item{Edwards 1971:} x <- sqrt(x); x <- x%*%t(x); x
<- 1/(x/number.loci); diag(x) <- 0; x <- sqrt(x)}
}
\examples{
# by pop:
y <- calc_genetic_distances(stickSNPs, facet = "pop", method = "Edwards"))
get.snpR.stats(y, "pop", "genetic_distance")

# by group and pop jointly
y <- calc_genetic_distances(stickSNPs, facet = "pop.group", method = "Edwards")
get.snpR.stats(y, "pop.group", "genetic_distance")

# by pop and fam seperately
y <- calc_genetic_distances(stickSNPs, facet = c("pop", "fam"), method = "Edwards")
get.snpR.stats(y, c("pop", "group"), "genetic_distance")

# individuals across all snps + plot
y <- calc_genetic_distances(stickSNPs)
heatmap(as.matrix(get.snpR.stats(y, type = "genetic_distance")$.base$.base$Edwards))

}
\references{
Edwards, A. W. F. (1971). Distances between populations on the
  basis of gene frequencies. Biometrics, 873-881.
}
\author{
William Hemstrom
}
