% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_het_hom_ratio}
\alias{calc_het_hom_ratio}
\title{Calculate the ratio of heterozygous/homozygous sites per individual.}
\usage{
calc_het_hom_ratio(x, facets = NULL)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{facets over which to split snps within samples. Takes only SNP
level facets. See \code{\link{Facets_in_snpR}} for details.}
}
\value{
A snpRdata object with heterozygote/homozygote ratios merged into the
  sample.stats slot.
}
\description{
Calculates the ratio of heterozygotes to homozygous sites across all SNPs
within each individual.
}
\examples{
# base facet
x <- calc_het_hom_ratio(stickSNPs)
get.snpR.stats(x, type = "sample")

# facet by chromosome
x <- calc_het_hom_ratio(stickSNPs, "group")
get.snpR.stats(x, "group", type = "sample")



}
\author{
William Hemstrom
}
