% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_het_hom_ratio}
\alias{calc_het_hom_ratio}
\title{Calculate the ratio of heterozygous/homozygous sites per individual.}
\usage{
calc_het_hom_ratio(x, facets = NULL, complex_averages = FALSE)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{facets over which to split snps within samples. Takes only SNP
level facets. See \code{\link{Facets_in_snpR}} for details.}

\item{complex_averages}{logical, default FALSE. If TRUE, will compute weighted averages for
complex (snp + sample metadata) facets. This can be quite time consuming, and so is generally
not recommended.}
}
\value{
A snpRdata object with heterozygote/homozygote ratios merged into the
  sample.stats slot.
}
\description{
Calculates the ratio of heterozygotes to homozygous sites across all SNPs
within each individual.
}
\examples{
# base facet
x <- calc_het_hom_ratio(stickSNPs)
get.snpR.stats(x, stats = "het_hom_ratio")

# facet by chromosome
x <- calc_het_hom_ratio(stickSNPs, "chr")
get.snpR.stats(x, "chr", stats = "het_hom_ratio")



}
\author{
William Hemstrom
}
