% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_Ho}
\alias{calc_Ho}
\title{Observed heterozygosity from SNP data.}
\usage{
calc_Ho(x, ecs, mDat = "NN", pop = NULL)
}
\arguments{
\item{x}{Input data, in "NN" or "0000" format, as given by format_snps output option 4 or 6.}

\item{ecs}{Number of extra metadata columns at the start of x.}

\item{pop}{FALSE or table, default FALSE. A table with population information for individuals. Individuals must be sorted in input data in the population order given in this table.}

\item{m.dat}{Character variable matching the coding for missing *genotypes* in x (typically "NN" or "0000").}
}
\value{
Data frame containing metadata, and Ho for each population in named columns.
}
\description{
\code{calc_Ho} calculates observed heterozygosity at each SNP, potentially in many populations.
}
\details{
Description of x:
   Contains metadata in columns 1:ecs. Remainder of columns contain genotype calls for each individual. Each row is a different SNP, as given by format_snps output options 4 or 6. If the data contains individuals from multiple populatoins, they must be in the same order as given to the pop agrument.
}
\examples{
#no seperate pops:
calc_Ho(stickSNPs, 3)

#seperate pops
pops <- table(substr(colnames(stickSNPs[,4:ncol(stickSNPs)]), 1, 3))
calc_Ho(stickSNPs, 3, pop = pops)

}
