% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_ne}
\alias{calc_ne}
\title{Caluclate effective population size.}
\usage{
calc_ne(
  x,
  facets = NULL,
  chr = NULL,
  NeEstimator_path = "/usr/bin/Ne2-1.exe",
  mating = "random",
  pcrit = c(0.05, 0.02, 0.01),
  methods = "LD",
  temporal_methods = c("Pollak", "Nei", "Jorde"),
  temporal_gens = NULL,
  max_ind_per_pop = NULL,
  outfile = "ne_out"
)
}
\arguments{
\item{x}{snpRdata object. The data for which Ne will be calculated.}

\item{facets}{character, default NULL. Categorical metadata variables by
which to break up analysis. See \code{\link{Facets_in_snpR}} for more
details. Only sample specific categories are allowed, all others will be
removed. If NULL, Ne will be calculated for all samples.}

\item{chr}{character, defualt NULL. An optional but recommended SNP specific
categorical metadata variable which designates chromosomes/linkage
groups/etc. Pairwise LD scores for SNPs with the same level of this
variable will be not be used to calculate Ne. Since physical linkage can
bias Ne estimtes, providing a value here is recommended.}

\item{NeEstimator_path}{character, default "/usr/bin/Ne2-1.exe". Path to the
NeEstimator executable.}

\item{mating}{character, default "random". The mating system to use. Options:
\itemize{ \item{"random"} Random mating. \item{"monogamy"} Monogomous
mating. }}

\item{pcrit}{numeric, default c(.05, .02, .01). Minimum minor allele
frequencies for which to calculate Ne. Rare alleles can bias esitmates, so
a range of values should be checked.}

\item{methods}{character, default "LD". LD estimation methods to use.
Options: \itemize{ \item{"LD"} Linkage Disequilibrium based estimation.
\item{"Ht"} Heterozygote excess. \item{"Coan"} Coancestry based.}}

\item{temporal_methods}{character, default c("Pollak", "Nei", "Jorde").
Methods to use for the temporal methods. See defaults for options. Not
currently supported.}

\item{temporal_gens}{data.frame, default NULL. Work in progress, not
currently supported.}

\item{max_ind_per_pop}{numeric, default NULL. Maximum number of individuals
to consider per population.}

\item{outfile}{character, default "ne_out". Prefix for output files. Note
that this function will return outputs, so there isn't a strong reason to
check this.}
}
\value{
A named list containing estimated Ne values (named "ne") and the
  original provided data, possibly with additional LD values (named "x").
}
\description{
Calculates effective population size for any given sample-level facets via
interface with the NeEstimator v2 program by Do et al. (2013).
}
\details{
Since physical linkage can cause mis-estimation of Ne, an optional snp-level
facet can be provided which designates chromosomes or linkage groups. Only
pairwise LD values between SNPs on different facet levels will be used.

Ne can be calculated via three different methods: \itemize{ \item{"LD"}
Linkage Disequilibrium based estimation. \item{"Ht"} Heterozygote excess.
\item{"Coan"} Coancestry based.} For details, please see the documentation
for NeEstimator v2.
}
\examples{
\dontrun{
# not run
# calculate Ne, noting not to use LD between SNPs on the same chromosome equivalent ("group") for every population.
ne <- calc_ne(stickSNPs, facets = "pop", chr = "group") # need to set the path argument to the local NeEstimator installation. 
get.snpR.stats(ne, "pop", type = "pop")}

}
\references{
Do C, Waples RS, Peel D, Macbeth GM, Tillett BJ, Ovenden JR. 2014
  NeEstimator v2: re-implementation of software for the estimation of
  contemporary effective population size (Ne) from genetic data. Mol. Ecol.
  Resour. 14, 209–214. (doi:10.1111/1755-0998.12157)
}
\author{
William Hemstrom
}
