% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_pairwise_fst}
\alias{calc_pairwise_fst}
\title{Pairwise FST from SNP data.}
\usage{
calc_pairwise_fst(
  x,
  facets,
  method = "WC",
  boot = FALSE,
  boot_par = FALSE,
  cleanup = TRUE
)
}
\arguments{
\item{x}{snpRdata. Input SNP data.}

\item{facets}{character. Categorical metadata variables by which to break up
analysis. See \code{\link{Facets_in_snpR}} for more details.}

\item{method}{character, default "wc". Defines the FST estimator to use.
Options: \itemize{ \item{wc: } Wier and Cockerham (1984).
\item{Genepop: } Rousset (2008), uses the genepop package. }}

\item{boot}{numeric or FALSE, default FALSE. The number of bootstraps to do.
See details.}

\item{boot_par}{numeric or FALSE, default FALSE. If a number, bootstraps will
be processed in parallel using the supplied number of cores.}

\item{cleanup}{logical, default TRUE. If TRUE, any new files created during
FST calculation will be automatically removed.}
}
\value{
A snpRdata object with pairwise FST as well as the number of total
 observations at each SNP in each comparison merged in to the pairwise.stats
 slot.
}
\description{
\code{calc_pairwise_fst} calculates pairwise FST for each SNP for each
possible pairwise combination of populations.
}
\details{
Calculates FST according to either Wier and Cockerham 1984 or using the
\code{\link[genepop]{Fst}} function from the genepop package (see references).

If the genpop option is used, several intermediate files will be created in
the default temporary directory (see \code{\link{tempfile}}).

The Wier and Cockerham (1984) and genepop methods tend to
produce very similar results both per SNP and per population.
Generally, the former option may be preferred for computational efficiency.

P-values for group level comparisons can be calculated via bootstrapping
using the boot option. Bootstraps are performed via randomly mixing
individuals amongst different levels of the supplied facet, and thus
the null hypothesis is that all groups are panmictic. P-values are calculated
according to \code{\link[ade4]{randtest}}, although that function is not
directly called. 

The data can be broken up categorically by either SNP and/or sample metadata,
as described in \code{\link{Facets_in_snpR}}. Since this is a pairwise
statistic, at least a single sample level facet must be provided.

Method Options: \itemize{ \item{"wc": }{Wier and Cockerham 1984.}
item{"Genepop": }{As used in genepop, Rousset
2008.}}
}
\examples{
# Using Wier and Cockerham 1984's method
x <- calc_pairwise_fst(stickSNPs, "pop")
get.snpR.stats(x, "pop", "fst")

\dontrun{
# Using genepop
x <- calc_pairwise_fst(stickSNPs, "pop", "genepop")
get.snpR.stats(x, "pop", "fst")

# bootstrap p-values for overall pairwise-Fst values
x <- calc_pairwise_fst(stickSNPs, "pop", boot = 5)
get.snpR.stats(x, "pop", "fst")
}
}
\references{
Wier and Cockerham (1984). \emph{Evolution}

Wier (1990). Genetic data analysis. Sinauer,  Sunderland, MA

Rousset (2008). \emph{Molecular Ecology Resources}
}
\author{
William Hemstrom
}
