% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_pairwise_ld}
\alias{calc_pairwise_ld}
\title{Pairwise LD from SNP data.}
\usage{
calc_pairwise_ld(x, facets = NULL, subfacets = NULL, ss = FALSE,
  par = FALSE, sr = FALSE)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{Facets to split by. Multi-level facets can be noted with a ".". "pop.family.chr", for example, will split by chr within each family and within each group/family level combo.}

\item{subfacets}{Subsets the facet levels to run. Given as a named list: list(fam = A) will run only fam A, list(fam = c("A", "B"), chr = 1) will run only fams A and B on chromosome 1. list(fam = "A", pop = "ASP") will run samples in either fam A or pop ASP, list(fam.pop = "A.ASP") will run only samples in fam A and pop ASP.}

\item{ss}{number of snps to subsample.}

\item{par}{number of parallel cores to run}

\item{sr}{should reports be supressed?}
}
\value{
Matrices containing rsq, Dprime, and p values for each SNP vs every other SNP. Can also produce a proximity table, which contains the rsq, Dprime, and p value for each pairwise comparison and the distance between the SNPs in those comparisons. Returns matrices and prox table as sequential elements in a named list.
}
\description{
\code{LD_full_pairwise} calculates LD between each pair of SNPs. If called as is, assumes one linkage group/chromosome and one population!
}
\details{
Matrix outputs are properly formated for LD_pairwise_heatmap.

Description of x:
   Contains metadata in columns 1:ecs. Remainder of columns contain genotype calls for each individual. Each row is a different SNP, as given by format_snps output options 4 or 6. Requires the column containing the position of the loci in base pairs be named "position". Note that this \emph{ignores populations}, split data into distinct populations before running.
}
\examples{
#returns prox table and LD matrices.
LD_full_pairwise(stickSNPs[stickSNPs$group == "groupI",1:53], ecs = 3)

}
