% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_pi}
\alias{calc_pi}
\title{Calculate PI from SNP data.}
\usage{
calc_pi(x, ecs)
}
\arguments{
\item{x}{Input SNP data, in allele count format as given by format_snps output option.}

\item{ecs}{Numeric, number of metadata columns at the start of x to keep for output.}
}
\value{
A vector of PI values, in the same order as the input SNPs.
}
\description{
\code{calc_pi} Calculates pi (genetic diversity/average number of pairwise differences) according to Hohenlohe et al. (2010) from SNP data. Returns a vector of pi values for each SNP as sorted in input.
}
\details{
Description of x:
   Must contain colums containing the number of *unique* alleles, total count of alleles sequenced in all individuals, and subsequent alleles counts for each observed allele in columns named "n_alleles", "n_total", "ni1", and "ni2". This matches the allele count/bayescan format as given by format_snps option one. Calculates pi for each row of data, and can therefore also contain a "pop" column and any other metadata columns such as SNP position.
}
\examples{
#get pop info, convert character data into the appropriate format, calculate pi, and add metadata.
pops <- table(substr(colnames(stickSNPs[,4:ncol(stickSNPs)]), 1, 3))
l <- list(c(names(pops)), as.numeric(pops))
ac <- format_snps(stickSNPs, 3, pop = l)
pi <- calc_pi(ac)
pi <- cbind(ac[,1:4], pi)

}
