% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{calc_prop_poly}
\alias{calc_prop_poly}
\title{Calculate the proportion of polymorphic loci.}
\usage{
calc_prop_poly(x, facets = NULL)
}
\arguments{
\item{x}{snpRdata object}

\item{facets}{facets over which to check for polymorphic loci. Can be any
combination of SNP/sample facets.
See \code{\link{Facets_in_snpR}} for details.}
}
\value{
A snpRdata object with prop_poly merged into the weighted.means
slot.
}
\description{
Calculates the proportion of polymorphic (genotyped) loci for any combination
of SNP and sample facets. Note: per window counts of polymorphic loci can be
calculated using \code{\link{calc_tajimas_d}}.
}
\details{
Note that per window counts of polymorphic loci can be calculated using
\code{\link{calc_tajimas_d}} (contained in the \code{num_seg} column).
Dividing this by the \code{n_snps} column will return the proportion of
polymorphic loci in each window if that is desired!
}
\examples{
# base facet
x <- calc_prop_poly(stickSNPs)
get.snpR.stats(x, stats = "prop_poly")$weighted.means

# multiple facets
x <- calc_prop_poly(stickSNPs, c("pop", "fam", "pop.fam", 
                                 "pop.fam.chr", "chr"))
get.snpR.stats(x, c("pop", "fam", "pop.fam", "pop.fam.chr", "chr"), 
               stats = "prop_poly")$weighted.means

}
\author{
William Hemstrom
}
