% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{check.installed}
\alias{check.installed}
\title{Utility to check if a required package is installed and install it if needed}
\usage{
check.installed(pkg, install.type = "basic")
}
\description{
Utility to check if a required package is installed and install it if needed
}
