% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata_methods.R, R/utility_functions.R
\docType{methods}
\name{extract_snpRdata}
\alias{extract_snpRdata}
\alias{genotypes}
\alias{genotypes<-}
\alias{snp.meta}
\alias{snp.meta<-}
\alias{sample.meta}
\alias{sample.meta<-}
\alias{genotypes,snpRdata-method}
\alias{genotypes<-,snpRdata-method}
\alias{snp.meta,snpRdata-method}
\alias{snp.meta<-,snpRdata-method}
\alias{sample.meta,snpRdata-method}
\alias{sample.meta<-,snpRdata-method}
\alias{[,snpRdata,ANY,ANY,ANY-method}
\alias{subset_snpR_data}
\alias{[}
\title{Get from or overwrite components of a snpRdata object}
\usage{
\S4method{genotypes}{snpRdata}(x)

\S4method{genotypes}{snpRdata}(x) <- value

\S4method{snp.meta}{snpRdata}(x, value)

\S4method{snp.meta}{snpRdata}(x) <- value

\S4method{sample.meta}{snpRdata}(x, value)

\S4method{sample.meta}{snpRdata}(x) <- value

\S4method{[}{snpRdata,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)

subset_snpR_data(x, .snps = 1:nsnps(x), .samps = 1:nsamps(x), ...)
}
\arguments{
\item{x}{snpRdata object.}

\item{value}{Genotypes, snp metadata, or sample metadata}

\item{i}{numeric, default \code{1:nrow(x)}. Row numbers corresponding to
SNPs to keep. Negative subscripts allowed.}

\item{j}{numeric, default \code{1:nrow(x)}. Row numbers corresponding to
SNPs to keep. Negative subscripts allowed.}

\item{...}{Facet subsetting may be specified by providing the facet as an 
argument and then providing the levels to keep or remove. Setting
pop = "ASP", for example, would keep only samples in the "ASP" level of the
pop facet, and setting pop.fam = "ASP.A" would keep only samples the ASP
pop and the A fam. Negative subscripts are allowed: pop = -c("ASP", "PAL")
would remove samples in the ASP or PAL pops. Subsetting by 
multiple facets is supported, although negative and positive subscripts
cannot be mixed across sample or SNP facets. They may be mixed between the
two.}

\item{drop}{logical, default FALSE. Deprecated.}

\item{.snps}{numeric, default \code{1:nrow(x)}. Row numbers corresponding to
SNPs to keep. Negative subscripts allowed.}

\item{.samps}{numeric, default \code{1:ncol(x)}. Column numbers corresponding
to samples to keep. Negative subscripts allowed.}
}
\description{
Fetch or overwirte the major parts of a snpRdata object (genotypes, snp meta,
or sample meta). If overwritten, any calculated stats will be removed, since
their values may be dependant upon changes in metadata.

Subsets snpRdata objects by specific snps, samples, facets, subfacets, ect.
Statistics will need to be recalculated after subsetting to avoid confusion. 
The bracket operators can be alternatively, following the same syntax.
}
\details{
If \emph{samples} are removed, most statistics will be thrown away, since
values like pi will change. In contrast, if \emph{snps} are removed, many
statistics will simply be subset. LD, bootstraps, and window stats will
\emph{always} be discarded, so overwrite with caution.

Sample and snp facets to subset over can be provided. Facet levels to keep are
provided in the corresponding subfacet arguments. Facets designated as
described in \code{\link{Facets_in_snpR}}.
}
\section{Functions}{
\itemize{
\item \code{genotypes,snpRdata-method}: view genotypes

\item \code{genotypes<-,snpRdata-method}: set genotypes

\item \code{snp.meta,snpRdata-method}: view snp meta

\item \code{snp.meta<-,snpRdata-method}: set snp meta

\item \code{sample.meta,snpRdata-method}: view sample meta

\item \code{sample.meta<-,snpRdata-method}: set sample meta

\item \code{[,snpRdata,ANY,ANY,ANY-method}: extraction operator

\item \code{subset_snpR_data}: subset_snpR_data
}}

\examples{
# copy test data
test <- stickSNPs

# show genotypes
genotypes(test)

# show or overwrite snp meta
snp.meta(test)
snp.meta(test) <- data.frame(pos = sample(10000, nrow(test), replace = TRUE), chr = sample(LETTERS[1:4], nrow(test), replace = TRUE))

#show or overwrite sample meta
sample.meta(test)
sample.meta(test) <- data.frame(fam = sample(LETTERS[1:4], ncol(test), replace = TRUE), pop = sample(LETTERS[5:8], ncol(test), replace = TRUE))
# Keep only individuals in the ASP and PAL populations 
# and on the LGIX or LGIV chromosome.
subset_snpR_data(stickSNPs, pop = c("ASP", "PAL"), 
                 group = c("groupIX", "groupIV"))
                 
# keep individuals/SNPs in the first 10 columns/rows.
subset_snpR_data(stickSNPs, 1:10, 1:10)

# negative subscripts: remove individuals in family B
subset_snpR_data(stickSNPs, fam = -"B")

# negative subscripts: remove individuals in pop PAL AND fam B or C, 
# and keep only SNPs on LGIV
subset_snpR_data(stickSNPs, pop.fam = -c("PAL.B", "PAL.C"),
                 group = "groupIV")
                 
# using the bracket operator, same as example 1:
stickSNPs[pop = c("ASP", "PAL"), group = c("groupIX", "groupIV")]
}
