% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata_methods.R
\docType{methods}
\name{extract_snpRdata}
\alias{extract_snpRdata}
\alias{genotypes}
\alias{genotypes<-}
\alias{snp.meta}
\alias{snp.meta<-}
\alias{sample.meta}
\alias{sample.meta<-}
\alias{genotypes,snpRdata-method}
\alias{genotypes<-,snpRdata-method}
\alias{snp.meta,snpRdata-method}
\alias{snp.meta<-,snpRdata-method}
\alias{sample.meta,snpRdata-method}
\alias{sample.meta<-,snpRdata-method}
\title{Get from or overwrite components of a snpRdata object}
\usage{
\S4method{genotypes}{snpRdata}(x)

\S4method{genotypes}{snpRdata}(x) <- value

\S4method{snp.meta}{snpRdata}(x, value)

\S4method{snp.meta}{snpRdata}(x) <- value

\S4method{sample.meta}{snpRdata}(x, value)

\S4method{sample.meta}{snpRdata}(x) <- value
}
\arguments{
\item{x}{snpRdata object to get genotype data from.}

\item{value}{Genotypes, snp metadata, or sample metadata}
}
\description{
Fetch or overwirte the major parts of a snpRdata object (genotypes, snp meta,
or sample meta). If overwritten, any calculated stats will be removed, since
their values may be dependant upon changes in metadata.
}
\section{Functions}{
\itemize{
\item \code{genotypes,snpRdata-method}: view genotypes

\item \code{genotypes<-,snpRdata-method}: set genotypes

\item \code{snp.meta,snpRdata-method}: view snp meta

\item \code{snp.meta<-,snpRdata-method}: set snp meta

\item \code{sample.meta,snpRdata-method}: view sample meta

\item \code{sample.meta<-,snpRdata-method}: set sample meta
}}

\examples{
# copy test data
test <- stickSNPs

# show genotypes
genotypes(test)

# show or overwrite snp meta
snp.meta(test)
snp.meta(test) <- data.frame(pos = sample(10000, nrow(test), replace = TRUE), chr = sample(LETTERS[1:4], nrow(test), replace = TRUE))

#show or overwrite sample meta
sample.meta(test)
sample.meta(test) <- data.frame(fam = sample(LETTERS[1:4], ncol(test), replace = TRUE), pop = sample(LETTERS[5:8], ncol(test), replace = TRUE))
}
