% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{filter_snps}
\alias{filter_snps}
\title{Filter SNPs in snpRdata objects.}
\usage{
filter_snps(x, maf = FALSE, hf_hets = FALSE, min_ind = FALSE,
  min_loci = FALSE, re_run = "partial", maf.facets = NULL,
  non_poly = TRUE, bi_al = TRUE)
}
\arguments{
\item{x}{snpRdata object.}

\item{maf}{FALSE or numeric between 0 and 1, default FALSE. Minimum
acceptable minor allele frequency}

\item{hf_hets}{FALSE or numeric between 0 and 1, default FALSE. Maximum
acceptable heterozygote frequency.}

\item{min_ind}{FALSE or integer, default FALSE. Minimum number of individuals
in which a loci must be sequenced.}

\item{min_loci}{FALSE or numeric between 0 and 1, default FALSE. Minimum
proportion of SNPs an individual must be genotyped at.}

\item{re_run}{FALSE, "partial", or "full", default "partial". How should loci
be re_filtered after individuals are filtered?}

\item{maf.facets}{FALSE or character, default FALSE. Sample-specific facets
over which the maf filter can be checked.}

\item{non_poly}{boolean, default TRUE. Should non-polymorphic loci be
removed?}

\item{bi_al}{boolean, default TRUE. Should non-biallelic SNPs be removed?}
}
\value{
A data.frame in the same format as the input, with SNPs and
  individuals not passing the filters removed.
}
\description{
\code{filter_snps} filters snpRdata objects to remove SNPs or individuals which fail
to pass user defined thresholds for several statistics. Since this function removes
all calculated statistics, etc. from the snpRdata object, this should usually be the first step
in an analysis. See details for filters.
}
\details{
Possible filters:
\itemize{
   \item{maf, minor allele frequency: }{removes SNPs where the minor allele frequency is too low. Can look for mafs below #'provided either globally or search each population individually.}
   \item{hf_hets, high observed heterozygosity: }{removes SNPs where the observed heterozygosity is too high.}
   \item{min_ind, minimum individuals: }{removes SNPs that were genotyped in too few individuals.}
   \item{min_loci, minimum loci: }{removes individuals sequenced at too few loci.}
   \item{non_poly, non-polymorphic SNPs: }{removes SNPs that are not polymorphic (not true SNPs).}
   \item{bi_al, non-biallelic SNPs: }{removes SNPs that have more than two observed alleles.}
}

Note that filtering out poorly sequenced individuals creates a possible
conflict with the loci filters, since after individuals are removed, some loci
may no longer pass filters. For example, if a portion of individuals in one
population all carry the only instances of a rare minor allele that still
passes the maf threshold, removing those individuals may cause the loci to no
longer be polymorphic in the sample.

To counter this, the "re_run" argument can be used to pass the data through a
second filtering step after individuals are removed. By default, the "partial"
re-run option is used, which re-runs only the non-polymorphic filter (if it
was originally set), since these may cause downstream analysis errors. The
"full" option re-runs all set filters. Note that re-running any of these
filters may cause individuals to fail the individual filter after loci
removal, and so subsequent tertiary re-running of the individual filters,
followed by the loci filters, and so on, could be justified. This function
stops after the second loci re-filtering, since that step is likely to be the
most important to prevent downstream analytical errors.

Via the "maf.filter" argument, this function can filter by minor allele
frequencies in either \emph{all} samples or \emph{each level of a supplied
sample specific facet and the entire dataset}. In the latter case, any SNPs
that pass the maf filter in \emph{any} facet level are considered to pass the
filter. The latter should be used in instances where populaiton sizes are very
different or there are \emph{many} populations, and thus common alleles of
interest in one population might be otherwise filtered out. With very small
populations, however, this may leave noise in the sample! In most cases,
filtering the entire dataset is sufficient. Facets should be provided as
described in \code{\link{Facets_in_snpR}}.
}
\examples{
# Filter with a minor allele frequency of 0.05, maximum heterozygote frequency
# of 0.55, 250 minimum individuals, and at least 75\% of loci sequenced per
# individual.
filter_snps(stickSNPs, maf = 0.05, hf_hets = 0.55, min_ind = 250, min_loci = .75)

# The same filters, but with minor allele frequency considered per-population
# and a full re-run of loci filters after individual removal.
filter_snps(stickSNPs, maf = 0.05, hf_hets = 0.55, min_ind = 250, min_loci = .75, re_run = "full", maf.facets = "pop")

}
