% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{filter_snps}
\alias{filter_snps}
\title{Filter SNPs in snpRdata objects.}
\usage{
filter_snps(
  x,
  maf = FALSE,
  hf_hets = FALSE,
  hwe = FALSE,
  fwe_method = "none",
  singletons = FALSE,
  min_ind = FALSE,
  min_loci = FALSE,
  re_run = "partial",
  maf_facets = NULL,
  hwe_facets = FALSE,
  non_poly = TRUE,
  bi_al = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{x}{snpRdata object.}

\item{maf}{numeric between 0 and 1 or FALSE, default FALSE. Minimum acceptable
minor allele frequency.}

\item{hf_hets}{numeric between 0 and 1 or FALSE, default FALSE. Maximum
acceptable heterozygote frequency.}

\item{hwe}{numeric between 0 and 1 or FALSE, default FALSE. Minimum acceptable
HWE p-value.}

\item{fwe_method}{character, default "none". Option to use for Family-Wise
Error rate correction for HWE filtering. If requested, only SNPs with
p-values below the alpha provided to the \code{hwe} argument \emph{after FWE
correction} will be removed. See \code{\link[stats]{p.adjust}} for
information on method options.}

\item{singletons}{logical, default FALSE. If TRUE, removes singletons (loci
where there is only a single minor allele). If population sizes are
reasonably high, this is more or less redundant if \code{maf} is also set.}

\item{min_ind}{numeric between 0 and 1 or FALSE, default FALSE. Minimum
proportion of individuals in which a loci must be sequenced.}

\item{min_loci}{numeric between 0 and 1 or FALSE, default FALSE. Minimum
proportion of SNPs at which an individual must be genotyped.}

\item{re_run}{character or FALSE, default "partial". When individuals are
removed via min_ind, it is possible that some SNPs that initially passed
filtering steps will now violate some filters. SNP filters can be re-run
automatically via several methods: \itemize{ \item{partial: } Re-filters for
non-polymorphic loci (non_poly) only, if that filter was requested
initially. \item{full: } Re-runs the full filtering scheme (save for
min_loci).}}

\item{maf_facets}{character or FALSE, default FALSE. Defines a sample facet
over which the minor allele frequency can be checked. SNPs will only fail the
maf filter if they fail in every level of every provided facet.}

\item{hwe_facets}{character or FALSE, default FALSE. Defines a sample facet
over which the hwe filter can be checked. SNPs will fail the hwe filter if
they fail in any level of any provided facet.}

\item{non_poly}{logical, default TRUE. If TRUE, non-polymorphic loci will be
removed.}

\item{bi_al}{logical, default TRUE. If TRUE, loci with more than two alleles
will be removed. Note that this is mostly an internal argument and should
rarely be used directly, since import.snpR.data and other snpRdata object
creation functions all pass SNPs through this filter because many snpR
functions will fail to work if there are more than two alleles at a locus.}

\item{verbose}{Logical, default TRUE. If TRUE, some progress updates and
filtering notes will be printed to the console.}
}
\value{
A data.frame in the same format as the input, with SNPs and
 individuals not passing the filters removed.
}
\description{
\code{filter_snps} filters snpRdata objects to remove SNPs or individuals
which fail to pass user defined thresholds for several statistics. Since this
function removes all calculated statistics, etc. from the snpRdata object,
this should usually be the first step in an analysis. See details for filters.
}
\details{
Possible filters: \itemize{ \item{maf, minor allele frequency: }{removes SNPs
where the minor allele frequency is too low. Can look for mafs below
#'provided either globally or search each population individually.}
\item{hf_hets, high observed heterozygosity: }{removes SNPs where the observed
heterozygosity is too high.} \item{min_ind, minimum individuals: }{removes
SNPs that were genotyped in too few individuals.} \item{min_loci, minimum
loci: }{removes individuals sequenced at too few loci.} \item{non_poly,
non-polymorphic SNPs: }{removes SNPs that are not polymorphic (not true
SNPs).} \item{bi_al, non-biallelic SNPs: }{removes SNPs that have more than
two observed alleles. This is mostly an internal argument, since the various
snpRdata import options use it automatically to prevent downstream errors in
other snpR functions. } }

Note that filtering out poorly sequenced individuals creates a possible
conflict with the loci filters, since after individuals are removed, some loci
may no longer pass filters. For example, if a portion of individuals in one
population all carry the only instances of a rare minor allele that still
passes the maf threshold, removing those individuals may cause the loci to no
longer be polymorphic in the sample.

To counter this, the "re_run" argument can be used to pass the data through a
second filtering step after individuals are removed. By default, the "partial"
re-run option is used, which re-runs only the non-polymorphic filter (if it
was originally set). The "full" option re-runs all set filters. Note that
re-running any of these filters may cause individuals to fail the individual
filter after loci removal, and so subsequent tertiary re-running of the
individual filters, followed by the loci filters, and so on, could be
justified. This is not done automatically here.

Via the "maf_facets" argument, this function can filter by minor allele
frequencies in either \emph{all} samples or \emph{each level of a supplied
sample specific facet and the entire dataset}. In the latter case, any SNPs
that pass the maf filter in \emph{any} facet level are considered to pass the
filter. The latter should be used in instances where population sizes are very
different, there are \emph{many} populations, and/or allele frequencies are
very different between populations and thus common alleles of interest in one
population might be otherwise filtered out.

The "hwe_facets" argument is the inverse of this: loci will be removed if they
fail the provided hwe filter in any facet level. In both cases, Facets should
be provided as described in \code{\link{Facets_in_snpR}}.
}
\examples{
# Filter with a minor allele frequency of 0.05, maximum heterozygote 
# frequency of 0.55, 50\% minimum individuals, and at least 75\% of loci 
# sequenced per individual.
filter_snps(stickSNPs, maf = 0.05, hf_hets = 0.55, 
            min_ind = 0.5, min_loci = 0.75)

# The same filters, but with minor allele frequency considered per-population
# and a full re-run of loci filters after individual removal.
filter_snps(stickSNPs, maf = 0.05, hf_hets = 0.55, min_ind = 0.5, 
            min_loci = 0.75, re_run = "full", maf_facets = "pop")

}
\author{
William Hemstrom
}
