% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{gap_snps}
\alias{gap_snps}
\title{Filter down to one SNPs every \emph{n} bases.}
\usage{
gap_snps(x, facet = NULL, n)
}
\arguments{
\item{x}{snpRdata object}

\item{facet}{character, default NULL. SNP facet specifying chromosomes or 
scaffolds. SNP positions will be independantly considered depending on the
facet level.}

\item{n}{Integer. Specifies the minium distance between selected SNPs.}
}
\value{
A snpRdata object containing the selected SNPs.
}
\description{
\code{rgap_snps} grabs one SNPs every \emph{n} bases. This can be
used to filter for SNPs in tight LD, such as when only one SNP per RAD-tag is
desired.
}
\details{
Note that this approach takes the first SNP every \emph{n} bases, and so is
non-random. \code{\link{filter_snps}} can be used beforehand to ensure that
the selected SNPs are above a desired quality threshold to ensure that poor
SNPs are not selected over loci with more robust genotyping data.

It is strongly recommended to provide a SNP facet to the facet argument. This
will define "chromosomes" for the purpose of selecting SNPs. If not set, snpR
will assume that \emph{all SNPs are on the same chromosome}, which may
produce undesired results.
}
\author{
William Hemstrom
}
