% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothing_functions.R
\name{gaussian_weight}
\alias{gaussian_weight}
\title{Calculate gaussian weights.}
\usage{
gaussian_weight(p, c, s)
}
\arguments{
\item{p}{numeric. Position of value to smooth.}

\item{c}{numeric. Position of center of window.}

\item{s}{numeric. Sigma, scaling factor.}
}
\value{
A numeric value, the relative weight of the point.
}
\description{
Get the gaussian weight of a statistic given its position relative to the
center and scaling factor sigma.
}
