% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{get.task.list}
\alias{get.task.list}
\title{List unique tasks/options for facets.
Internal function to get a list of tasks to run (one task per unique sample/snp level facet!). The source arguement specifies what kind of statistics are being grabbed.}
\usage{
get.task.list(x, facets, source = "stats")
}
\description{
List unique tasks/options for facets.
Internal function to get a list of tasks to run (one task per unique sample/snp level facet!). The source arguement specifies what kind of statistics are being grabbed.
}
