% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{import.snpR.data}
\alias{import.snpR.data}
\title{Import genotype and metadata into a snpRdata object.}
\usage{
import.snpR.data(genotypes, snp.meta, sample.meta, mDat)
}
\description{
\code{import.snpR.data} converts genotype and meta data to the snpRdata class, which stores raw genotype data, sample and locus specific metadata, useful data summaries, repeatedly internally used tables, calculated summary statistics, and smoothed statistic data.
}
\details{
The snpRdata class is built to contain SNP genotype data for use by functions in the snpR package. It inherits from the S3 class data.frame, in which the genotypes are stored, and can be manipulated identically. It also stores sample and locus specific metadata, genomic summary information, and any results from most snpR functions. The raw data for each of these latter objects is accessable via the at operator.
Genotypes are stored in the "character" format, as output by format_snps(). Missing data is noted with "NN".
}
