% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_functions.R
\name{LD_full_pairwise}
\alias{LD_full_pairwise}
\title{Pairwise LD from SNP data.}
\usage{
LD_full_pairwise(x, ecs, prox_table = TRUE, matrix_out = TRUE, mDat = "N")
}
\arguments{
\item{x}{Input data, in either numeric or character formats, as given by format_snps output options 2 or 6.}

\item{ecs}{Number of extra metadata columns at the start of x.}

\item{prox_table}{If TRUE, a proximity table is produced.}

\item{matrix_out}{If TRUE, pairwise LD matrices are produced.}

\item{mDat}{Character variable matching the coding for missing *alleles* in x (typically "N" or "00").#' @return Data frame containing metadata, and Ho for each population in named columns.}
}
\value{
Matrices containing rsq, Dprime, and p values for each SNP vs every other SNP. Can also produce a proximity table, which contains the rsq, Dprime, and p value for each pairwise comparison and the distance between the SNPs in those comparisons. Returns matrices and prox table as sequential elements in a named list.
}
\description{
\code{LD_full_pairwise} calculates LD between each pair of SNPs. If called as is, assumes one linkage group/chromosome and one population!
}
\details{
Matrix outputs are properly formated for LD_pairwise_heatmap.

Description of x:
   Contains metadata in columns 1:ecs. Remainder of columns contain genotype calls for each individual. Each row is a different SNP, as given by format_snps output options 4 or 6. Requires the column containing the position of the loci in base pairs be named "position". Note that this \emph{ignores populations and linkage group/chromosome}, split data into distinct populations and linkage groups or use Full_LD_w_groups before running.
}
\examples{
#returns prox table and LD matrices.
LD_full_pairwise(stickSNPs[stickSNPs$group == "groupI",1:53], ecs = 3)

}
