% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_clusters}
\alias{plot_clusters}
\title{tSNE from genetic data.}
\usage{
plot_clusters(x, facets = FALSE, plot_type = c("PCA", "tSNE"),
  check_duplicates = FALSE, minimum_percent_coverage = FALSE,
  interpolation_method = "bernoulli", dims = 2, initial_dims = 50,
  perplexity = FALSE, theta = 0, iter = 1000, ...)
}
\arguments{
\item{x}{Input presence/absence data, as described in details.}

\item{dims}{Integer, output dimensionality, default 2.}

\item{initial_dims}{Integer, default 50. The number of dimensions retained in the initial PCA step.}

\item{perplexity}{Perplexity parameter, by default found by \code{\link[mmtsne]{hbeta}}, with beta = 1.}

\item{theta}{Theta parameter from \code{\link[Rtsne]{Rtsne}}. Default 0, an exhaustive search.}

\item{iter}{Integer, default 5000. Number of tSNE iterations to perform.}

\item{...}{Other arguments, passed to \code{\link[Rtsne]{Rtsne}}.}

\item{ecs}{Numeric. The number of extra metadata columns at the start of the input. Must be more that two to avoid errors. I really should fix that at some point. Includes the "pop" collumn.}

\item{plot.vars}{FALSE or character vector, default "pop". If FALSE, a basic plot is produced. Up to two variables to be plotted with names corresponding to column names in x be given as a character vector.}

\item{c.dup}{boolean, default FALSE. Should duplicate individuals be searched for and removed? This is very slow if the data set is large!}

\item{mc}{Numeric or FALSE, default FALSE. Should poorly sequenced individuals be removed? If so, what is the minimum acceptable count of sequenced loci (as specificed in the vector provided to counts)?}

\item{counts}{Numeric vector, default FALSE, containing the number of loci sequenced per individual.}

\item{do.plot}{boolean, default TRUE. Should a plot be produced?}
}
\value{
A list containing the raw tSNE output and a tSNE plot in the form of a ggplot graphical object. The plot can be changed as usual with ggplot objects.
}
\description{
\code{tSNEfromPA} creates a ggplot object tSNE from presence/absense allelic data using the Barnes-Hut simulation at theta>0 implemented in \code{\link[Rtsne]{Rtsne}}. If individuals which were sequenced at too few loci are to be filtered out, both the mc and counts argument must be provided.
}
\details{
See the documentaion for \code{\link[Rtsne]{Rtsne}} for details. Defaults match those of \code{\link[Rtsne]{Rtsne}}. Argument documentation taken from that function.

Description of x:
   SNP or other allelic data in presence/absence format, as given by \code{\link{format_snps}} option 7. An additional column of population IDs titled "pop" must also be provided.

This function results from collaboration with Matt Thorstensen.
}
\examples{
tSNEfromPA(stickPA, 2, c.dup = TRUE)
}
