% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_pairwise_ld_heatmap}
\alias{plot_pairwise_ld_heatmap}
\title{Create a heatmap from pairwise linkage data.}
\usage{
plot_pairwise_ld_heatmap(
  x,
  facets = NULL,
  snp.subfacet = NULL,
  sample.subfacet = NULL,
  LD_measure = "CLD",
  r = NULL,
  l.text = "CLD",
  viridis.option = "inferno",
  gradient_colors = NULL,
  title = NULL,
  t.sizes = c(16, 13, 10, 12, 10),
  background = "white",
  simplify_output = FALSE
)
}
\arguments{
\item{x}{snpRdata object.}

\item{facets}{character, default NULL. Categorical metadata variables by which
to break up plots. Must match facets for which LD data has been previously
calculated, and only a single facet can be plotted at once. See
\code{\link{Facets_in_snpR}} for more details.}

\item{snp.subfacet}{character, default NULL. Specific snp-specific levels of
the provided facet to plot. See examples.}

\item{sample.subfacet}{character, default NULL. Specific sample-specific
levels of the provided facet to plot. See examples.}

\item{LD_measure}{character, default rsq. LD metric to plot. Must be present
in the calculated LD data.}

\item{r}{Numeric. Region of the chromosome to subset and plot. Given in kb in
the format numeric vector c(lower, upper).}

\item{l.text}{character, default "CLD". Legend title.}

\item{viridis.option}{character, default "inferno". Viridis color scale option
to use. Other color scales may be substituted by appending the
scale_color_continuous and scale_fill_continuous ggplot functions to the
produced plot using the '+' operator. See
\code{\link[ggplot2]{scale_gradient}} for details.}

\item{gradient_colors}{character vector of length 2 or NULL, default NULL. If
provided, the low and high LD colors for the color scale. Provided as an
alternative to the viridis scales. Take care to set a \code{background}
color not on this scale. For example, \code{c("white", "black")} is an
excellent choice, but will make missing data inaccurately appear  appear as
if it has a very low LD unless a different \code{background} color is
selected.}

\item{title}{character. Plot title.}

\item{t.sizes}{numeric, default c(16, 13, 10, 12, 10). Text sizes, given as
c(title, legend.title, legend.ticks, axis, axis.ticks).}

\item{background}{character, default "white". Background color for plot.}

\item{simplify_output}{If TRUE, only the ggplot object will be return. This
is optimal, since the data is already returned in that object, but is not
the default due to backwards consistency with old code.}
}
\value{
A list containing: \itemize{ \item{plot: } A pairwise LD heatmap as a
 ggplot object. \item{dat: } Data used to generate the ggplot object. } If
 \code{simplify_output} is \code{FALSE}, only the ggplot object is returned.
}
\description{
Prepares a ggplot2 heatmap from pairwise linkage disequilibrium data stored in
a snpRdata object.
}
\details{
Since the output is a ggplot object, options can be added or changed by adding
"+ function()" to the end of \code{LD_pairwise_heatmap}. Some common options
are also built into this function as arguments, but can be overwritten freely.

Specific facets and facet levels can be requested as long as LD data for those
facets has been previously calculated. Facets should be provided as described
in \code{\link{Facets_in_snpR}}. Only one facet can be plotted at once. Plots
for individual SNP or sample-specific facet levels for the specified facet can
also be requested. See examples. For facets with many categories, this is
strongly recommended, since plots with many facet levels become
computationally challenging to produce and difficult to interpret. A facets
argument of NULL will plot the base level facet, a facet argument of "all" is
not accepted.

Note that NA LD values will be colored white in the resulting plot.
}
\examples{
\dontrun{
# get LD data
dat <- calc_pairwise_ld(stickSNPs, c("pop.chr"))

# produce plots for linkage group IX in the ASP and CLF populations.
plot_pairwise_ld_heatmap(dat, c("pop.chr"), "groupIX", c("ASP", "CLF"))

# produce plots for every population for linkage group IV
plot_pairwise_ld_heatmap(dat, c("pop.chr"), "groupIV")
}

}
\author{
William Hemstrom

Nicholas Sard
}
