% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_structure}
\alias{plot_structure}
\title{Create STRUCTURE-like cluster plots}
\usage{
plot_structure(
  x,
  facet = NULL,
  facet.order = NULL,
  k = 2,
  method = "snmf",
  reps = 1,
  iterations = 1000,
  I = NULL,
  alpha = 10,
  qsort = "last",
  qsort_K = "last",
  clumpp = T,
  clumpp_path = "/usr/bin/CLUMPP.exe",
  clumpp.opt = "greedy",
  ID = NULL,
  viridis.option = "viridis",
  alt.palette = NULL,
  t.sizes = c(12, 12, 12),
  ...
)
}
\arguments{
\item{x}{snpRdata object, list of Q matrices (sorted by K in the first level
and run in the second), or a character string designating a pattern
matching Q matrix files in the current working directories.}

\item{facet}{character, default NULL. If provided, individuals will not be
noted on the x axis. Instead, the levels of the facet will be noted. Only a
single, simple, sample specific facet may be provided. Individuals must be
sorted by this facet in x. If Q matrices are provided (either directly or
via file path), this should instead be a vector of group identities for
each individual (populations, etc.).}

\item{facet.order}{character, default NULL. Optional order in which the
levels of the provided facet should appear on the plot, left to right.}

\item{k}{numeric, default 2. The maximum of k (number of clusters) for which
to run the clustering/assignment algorithm. The values 2:k will be run.}

\item{method}{character, default "snmf". The clustering/assignment method to
run. Options: \itemize{\item{snmf: } sNMF (sparse Non-Negative Matrix
Factorization). \item{snapclust: } Maximum-likelihood genetic clustering.}
See \code{\link[LEA]{main_sNMF}} or
\code{\link[adegenet]{snapclust.choose.k}} for details, respectively.}

\item{reps}{numeric, default 1. The number of independent clustering
repititions to run.}

\item{iterations}{numeric or Inf, default 1000. For snapclust, the maximum
number of iterations to run.}

\item{I}{numeric or NULL, default NULL. For snmf, how many SNPs should be
used to initialize the search? Initializing with a subset of the total SNPs
can radically speed up computation time for large datasets.}

\item{alpha}{numeric, default 10. For sNMF, determines the regularization
parameter. For small datasets, this can have a large effect, and should
probably be larger than the default. See documentation for
\code{\link[LEA]{main_sNMF}}.}

\item{qsort}{character, numeric, or FALSE, default "last". Determines if
individuals should be sorted (possibly within facet levels) by cluster
assignment proportion. If not FALSE, determines which cluster to use for
sorting (1:k). If "last" or "first" sorts by those clusters.}

\item{qsort_K}{numeric or character, default "last". If qsorting is
performed, determines the reference k value by which individuals are
sorted. If "first" or "last", sorts by k = 2 or k = k, respectively.}

\item{clumpp}{logical, default T. Specifies if CUMPP should be run to
collapse results across multiple reps. If FALSE, will use only the first
rep for plotting.}

\item{clumpp_path}{character, default "/usr/bin/CLUMPP.exe". Path to the
clumpp executable, required if clumpp = T.}

\item{clumpp.opt}{character, default "greedy". Designates the CLUMPP method
to use. Options: \itemize{ \item{fullsearch: } Search all possible
configurations. Slow. \item{greedy: } The standard approach. Slow for large
datasets at high k values. \item{large.k.greedy: } A fast but less accurate
approach. } See CLUMPP documentation for details.}

\item{ID}{character or NULL, default NULL. Designates a column in the sample
metadata containing sample IDs.}

\item{viridis.option}{character, default "viridis". Viridis color scale
option. See \code{\link[ggplot2]{scale_gradient}} for details.}

\item{alt.palette}{charcter or NULL, default NULL. Optional palette of colors
to use instead of the viridis palette.}

\item{t.sizes}{numeric, default c(12, 12, 12). Text sizes, given as
c(strip.title, axis, axis.ticks).}

\item{...}{additional arguments passed to either \code{\link[LEA]{main_sNMF}}
or \code{\link[adegenet]{snapclust.choose.k}}.}
}
\value{
A list containing: \itemize{\item{plot: } A ggplot object.
  \item{data: } A nested list of the raw Q matrices, organized by K and then
  by run. \item{plot_data: } The raw data used in constructing the ggplot.
  \item{K_plot: } A data.frame containing the value suggested for use in K
  selection vs K value for the selected method.}
}
\description{
Creates ggplot-based stacked barcharts of assignment probabilities (Q) into
an arbitrary 'k' number of clusters like those produced by the program
STRUCTURE. Runs for each value of k between 2 and the given number.
}
\details{
Individual cluster assignment probabilities can be calculated using several
different methods: \itemize{\item{snmf: } sNMF (sparse Non-Negative Matrix
Factorization). \item{snapclust: } Maximum-likelihood genetic clustering.}
These methods are not re-implemented in R, instead, this function calls the
\code{\link[LEA]{main_sNMF}} and \code{\link[adegenet]{snapclust.choose.k}}
functions instead. Please cite the references noted in those functions if
using this function. For snapclust, the "ward" method is used to initialize
clusters if one rep is requested, otherwise the clusters are started randomly
each rep. Other methods can be used by providing pop.ini as an additional
argument as long as only one rep is requested.

Multiple different runs can be conducted using the 'reps' argument, and the
results can be combined for plotting across all of these reps using the
clumpp option. This option calls the CLUMPP software package in order to
combine proportion population membership across multiple runs via
\code{\link[pophelper]{clumppExport}}. Again, please cite both CLUMPP and
pophelper if using this option.

Since CLUMPP is run independantly for each value of K, cluster identites
often "flip" between K values. For example individuals that are grouped into
cluster 1 and K = 3 may be grouped into cluster 2 at K = 4. To adjust this,
cluster IDs are iteratively adjusted across K values by flipping IDs such
that the euclidian distances between clusters at K and K - 1 are minimized.
This tends to produce consistant cluster IDs across multiple runs of K.

Individuals can be sorted into by membership proportion into different
clusters within populations using the qsort option.

Since the clustering and CLUMPP processes can be time consuming and outside
tools (such as NGSadmix or fastSTRUCTURE) may be prefered, a nested list of Q
matrices, sorted by K and then rep or a character string giving a pattern
matching saved Q matrix files in the current working directory may provided
directly instead of a snpRdata object. Note that the output for this
funciton, if run on a snpRdata object, will return a properly formatted list
of Q files (named 'data') in addition to the plot and plot data. This allows
for the plot to be quickly re-constructed using different sorting parameters
or facets. In these cases, the facet argument should instead be a vector of
group identifications per individuals.

Note that several files will be created in the working directory when using
this function that are not automatically cleaned after use.
}
\examples{
# basic sNMF
plot_structure(stickSNPs, "pop")

}
\references{
Frichot E, Mathieu F, Trouillon T, Bouchard G, Francois O.
  (2014). Fast and Efficient Estimation of Individual Ancestry Coefficients.
  \emph{Genetics}, 194(4): 973–983.

Frichot, Eric, and Olivier François (2015). LEA: an R package for
  landscape and ecological association studies. \emph{Methods in Ecology and
  Evolution}, 6(8): 925-929.

Beugin, M. P., Gayet, T., Pontier, D., Devillard, S., & Jombart,
  T. (2018). A fast likelihood solution to the genetic clustering problem.
  \emph{Methods in ecology and evolution}, 9(4), 1006-1016.

Francis, R. M. (2017). pophelper: an R package and web app to
  analyse and visualize population structure. \emph{Molecular ecology
  resources}, 17(1), 27-32.

Jakobsson, M., & Rosenberg, N. A. (2007). CLUMPP: a cluster
  matching and permutation program for dealing with label switching and
  multimodality in analysis of population structure. \emph{Bioinformatics},
  23(14), 1801-1806.
}
\author{
William Hemstrom
}
