% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_non_biallelic}
\alias{read_non_biallelic}
\title{Read in non-biallelic genotype data.}
\usage{
read_non_biallelic(
  genotypes,
  snp.meta = NULL,
  sample.meta = NULL,
  header_cols = 0,
  mDat = "0000",
  verbose = FALSE,
  .pass_filters = FALSE
)
}
\arguments{
\item{genotypes}{data.frame, \code{\link[vcfR]{vcfR}}, or a file path to either
vcf or tab-delimited file containing raw genotypes with SNPs are in rows and
individual samples are in columns. If raw genotypes, each genotype must be
stored in one cell (as in "001002" or "0102") with a matching length provided
to the "mDat" argument.}

\item{snp.meta}{data.frame, default NULL. Metadata for each SNP, must have a
number of rows equal to the number of SNPs in the dataset. If NULL, a single
"snpID" column will be added.}

\item{sample.meta}{data.frame, default NULL. Metadata for each individual
sample, must have a number of rows equal to the number of samples in the
dataset. If NULL, a single "sampID" column will be added.}

\item{header_cols}{numeric, default 0. Number of header columns containing
SNP metadata. Used if a tab delimited or STRUCTURE input file is
provided.}

\item{mDat}{character, default "000000", matching the encoding of missing
\emph{genotypes} in the data provided to the genotypes argument.}

\item{verbose}{Logical, default FALSE. If TRUE, will print a few status
updates and checks.}

\item{.pass_filters}{Internal, probably not for user use. Used to pass
filtering history when sub-setting when this function is called internally.}

\item{.skip_filters}{Internal, probably not for user use. Used to skip
re-filtering during sub-setting when this function is called internally.}
}
\description{
Reads in non-biallelic data stored as a vcf or tab-delimited data.
Currently only supports these data types.
}
