% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrapping_functions.R
\name{resample_long}
\alias{resample_long}
\title{Bootstrapped smoothed values.}
\usage{
resample_long(x, statistic, boots, sigma = 150, nk_weight = FALSE,
  fws = NULL, n_snps = T, report = 10000)
}
\arguments{
\item{x}{Input SNP data.}

\item{boots}{Number of bootstraps to run.}

\item{sigma}{Size variable in kb for gaussian smoothing. Full window size is 6*sigma.}

\item{nk_weight}{If true, weights smoothing by randomly drawing variables from a column titled "nk" in x.}

\item{fws}{If using a window with a fixed slide rather than snp centralized, provide a data frame of window positions.}

\item{n_snps}{If true, draws all random numbers up front. A column containing the number of snps in each selected window must be provided, titled "n_snps", in either x or fws. This allows for faster processing. Otherwise, random numbers will be drawn after this value is determined for each bootstrap.}

\item{report}{Progress report interval, in bootstraps.}
}
\value{
A numeric vector containing bootstrapped smoothed values.
}
\description{
\code{resample_long} creates a distribution of bootstrapped smoothed values for a desired statistic, as described by Hohenlohe et al. (2010), Population genomics of parallel adaptation in threespine stickleback using sequenced RAD tags. To run in parallel across multiple linkage groups/chromosomes, use  \code{\link{resample_long_par}}.
}
\examples{
resample_long(randPI[randPI$pop == "A" & randPI$group == "chr1",], "pi", 100, 200, TRUE, randSMOOTHed[randSMOOTHed$pop == "A" & randSMOOTHed$group == "chr1",], TRUE, 10)

}
