% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snpRdata_methods.R
\name{snpRdata_dims}
\alias{snpRdata_dims}
\alias{nsnps}
\alias{nsamps}
\alias{nrow}
\alias{ncol}
\alias{dim}
\alias{nsnps,snpRdata-method}
\alias{nrow,snpRdata-method}
\alias{nsamps,snpRdata-method}
\alias{ncol,snpRdata-method}
\alias{dim,snpRdata-method}
\title{Get the dimensions of a snpRdata object}
\usage{
\S4method{nsnps}{snpRdata}(x)

\S4method{nrow}{snpRdata}(x)

\S4method{nsamps}{snpRdata}(x)

\S4method{ncol}{snpRdata}(x)

\S4method{dim}{snpRdata}(x)
}
\arguments{
\item{x}{snpRdata object}
}
\description{
Get the dimensions of a snpRdata object
}
\section{Functions}{
\itemize{
\item \code{nsnps(snpRdata)}: get the number of SNPs

\item \code{nrow(snpRdata)}: identical to nsnps

\item \code{nsamps(snpRdata)}: get the number of samples

\item \code{ncol(snpRdata)}: identical to nsamps

\item \code{dim(snpRdata)}: get the number of SNPs and samples

}}
