% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{steelMSATs}
\alias{steelMSATs}
\title{Siletz Steelhead microsatellite snpRdata}
\format{
An object of class \code{snpRdata} with 13 rows and 1573 columns.
}
\usage{
steelMSATs
}
\description{
Contains example data for 13 microsatellite loci from 1573 steelhead from the
Siletz river, Oregon, USA. Samples come from both the summer and winter run
wild and hatchery populations. For more details, see Hemstrom et al 2018.
}
\details{
Sample metadata can be found using the \code{sample.meta} function.
Individuals have been randomly sorted into "families" for example purposes.
}
\keyword{datasets}
