% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\docType{data}
\name{steelRAW}
\alias{steelRAW}
\title{Contains example data for 13 microsatellite loci from 1573 steelhead from the
Siletz river, Oregon, USA. Samples come from both the summer and winter run
wild and hatchery populations. For more details, see Hemstrom et al 2018.}
\format{
An object of class \code{data.frame} with 1573 rows and 14 columns.
}
\usage{
steelRAW
}
\description{
The first column contains population info, each remaining column contains
information on one microsatellite. Note that this data is transposed with
respect to the SNP example data.
}
\details{
Individual genotypes are coded as "0000", where the first two numbers are an
identifer for the first allele and the third and fourth identify the second
allele.
}
\keyword{datasets}
