% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{tabulate_genotypes}
\alias{tabulate_genotypes}
\title{Tabulate allele and genotype counts at each locus.}
\usage{
tabulate_genotypes(x, mDat, verbose = F)
}
\arguments{
\item{x}{Input raw genotype data, where columns are individuals and rows are
snps. No metadata.}

\item{mDat}{Character. How are missing \emph{genotypes} noted?}

\item{verbose}{Logical. Should the function report progress?}
}
\value{
A list of matrices. gs is the genotype matrix, as is the allele
 matrix, and wm is the genotype matrix with missing genotypes.
}
\description{
\code{tabulate_genotypes} creates matricies containing counts of observed
alleles and genotypes at each locus.
}
\details{
This function is pirmarily used interally in several other funcitons.
}
\author{
William Hemstrom
}
