% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{update_calced_stats}
\alias{update_calced_stats}
\title{Update list of calculated stats for a vector of facet names}
\usage{
update_calced_stats(x, facets, stats)
}
\arguments{
\item{x}{snpRdata object to update}

\item{stats}{character. Name of the facet to update as calculated.}

\item{facet}{character. Facets to update, see \code{\link{Facets_in_snpR}}}
}
\value{
A snpRdata object identical to x but with calced stats updated.
}
\description{
Update list of calculated stats for a vector of facet names
}
\author{
William Hemstrom
}
