% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony_interface.R
\name{write_colony_input}
\alias{write_colony_input}
\title{Write an input for the COLONY pedigree assignment program.}
\usage{
write_colony_input(x, outfile = "colony_input", method = "FPLS",
  run_length = 2, sampleIDs = NULL, sibship_prior = 0,
  paternal_sib_size = NULL, maternal_sib_size = NULL, nruns = 1,
  seed = NULL, maternal_genotypes = NULL, paternal_genotypes = NULL,
  maternal_inclusion_prob = 0, paternal_inclusion_prob = 0,
  update_af = TRUE, dioecious = TRUE, inbreeding = TRUE,
  male_monogamous = FALSE, female_monogamous = FALSE,
  clone_inference = FALSE, sibship_scaling = TRUE, known_af = FALSE,
  precision = 2, dropout = 0.01, genotyping_error = 0.01,
  known_maternal_dryads = NULL, known_paternal_dryads = NULL,
  known_maternal_max_mismatches = 0, known_paternal_max_mismatches = 0,
  known_maternal_sibships = NULL, known_paternal_sibships = NULL,
  maternal_exclusions = NULL, paternal_exclusions = NULL,
  excluded_maternal_siblings = NULL, excluded_paternal_siblings = NULL)
}
\arguments{
\item{x}{snpRdata object containing offspring genotypes.}

\item{outfile}{character, default "colony_input". Output file name. A file
path may be provided (e.g. "colony/colony_run_1.txt").}

\item{method}{character, default "FPLS". Pedigree reconstruction method.
Options: \itemize{\item{"FLPS": }{Pure pairwise likelihood method, combines the full likelihood and pairwise likelihood methods. A good compromise between speed and accuracy.}
\item{"PLS": }{} }}

\item{run_length}{numeric in 1:4, default 2. Length of run, short/medium/long/verylong.}

\item{sibship_prior}{numeric in c(0, 0.25, 0.5, 1), default 0. Strength the sibship size prior (no prior, weak, medium, or strong). Values other than 0 require}

\item{sampleIDS}{character, default NULL. Name of a column in the sample metadata that designates sample identifications/"names". Each name must be unique!}

\item{sampleIDS}{character, default NULL. Name of a column in the sample metadata that designates sample identifications/"names". Each name must be unique!}
}
\description{
Creates an input file in the format required for command-line usage of the
COLONY pedigree program. Requires a snpRdata object containing offspring
genotypes and can optionally take a snpRdata object containing maternal or
paternal genotypes, or both.
}
