% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colony_interface.R
\name{write_colony_input}
\alias{write_colony_input}
\title{Write an input for the COLONY pedigree assignment program.}
\usage{
write_colony_input(x, outfile = "colony_input", method = "FPLS",
  run_length = 2, sampleIDs = NULL, sibship_prior = 0,
  paternal_sib_size = NULL, maternal_sib_size = NULL, nruns = 1,
  seed = NULL, maternal_genotypes = NULL, paternal_genotypes = NULL,
  maternal_inclusion_prob = 0, paternal_inclusion_prob = 0,
  update_af = TRUE, dioecious = TRUE, inbreeding = TRUE,
  male_monogamous = FALSE, female_monogamous = FALSE,
  clone_inference = FALSE, sibship_scaling = TRUE, known_af = FALSE,
  precision = 2, dropout = 0.01, genotyping_error = 0.01,
  known_maternal_dryads = NULL, known_paternal_dryads = NULL,
  known_maternal_max_mismatches = 0, known_paternal_max_mismatches = 0,
  known_maternal_sibships = NULL, known_paternal_sibships = NULL,
  maternal_exclusions = NULL, paternal_exclusions = NULL,
  excluded_maternal_siblings = NULL, excluded_paternal_siblings = NULL)
}
\description{
Write an input for the COLONY pedigree assignment program.
}
